refactor(torghut): split whitepaper autoresearch runner ownership
