+++ services/torghut/scripts/run_whitepaper_autoresearch_profit_target.py
@@ -122,6 +122,121 @@ import scripts.local_intraday_tsmom_replay as replay_mod
 import scripts.materialize_replay_tape as replay_materializer
 import scripts.run_strategy_factory_v2 as strategy_factory_runner
 
+from scripts.whitepaper_autoresearch_runner.candidate_board_fields import (
+    _candidate_board_score_rows as _candidate_board_score_rows,
+    _candidate_board_decimal_field as _candidate_board_decimal_field,
+    _candidate_board_int_field as _candidate_board_int_field,
+    _candidate_board_first_int_field as _candidate_board_first_int_field,
+)
+from scripts.whitepaper_autoresearch_runner.candidate_board_paper_probation import (
+    _paper_probation_candidate_payload as _paper_probation_candidate_payload,
+    _candidate_board_paper_probation_candidates as _candidate_board_paper_probation_candidates,
+    _candidate_board_paper_probation_candidate as _candidate_board_paper_probation_candidate,
+    _candidate_board_status_digest as _candidate_board_status_digest,
+    _candidate_board_double_oos_summary as _candidate_board_double_oos_summary,
+    _candidate_board_portfolio_promotion_subject as _candidate_board_portfolio_promotion_subject,
+)
+from scripts.whitepaper_autoresearch_runner.candidate_board_payloads import (
+    _candidate_board_factor_acceptance_summary as _candidate_board_factor_acceptance_summary,
+    _candidate_board_payload as _candidate_board_payload,
+    _paper_probation_handoff_payload as _paper_probation_handoff_payload,
+    _portfolio_with_runtime_closure_proof as _portfolio_with_runtime_closure_proof,
+    _runtime_closure_program_for_candidate as _runtime_closure_program_for_candidate,
+)
+from scripts.whitepaper_autoresearch_runner.candidate_board_runtime_windows import (
+    _candidate_board_hypothesis_manifest_ref as _candidate_board_hypothesis_manifest_ref,
+    _candidate_board_runtime_window_bounds as _candidate_board_runtime_window_bounds,
+    _candidate_board_date_only as _candidate_board_date_only,
+    _candidate_board_regular_session_bound as _candidate_board_regular_session_bound,
+    _candidate_board_runtime_window_import_bounds as _candidate_board_runtime_window_import_bounds,
+    _candidate_board_exact_replay_ledger_refs as _candidate_board_exact_replay_ledger_refs,
+    _candidate_board_runtime_import_args as _candidate_board_runtime_import_args,
+    _candidate_board_runtime_window_import_plan as _candidate_board_runtime_window_import_plan,
+    _candidate_factor_acceptance_replay_metadata as _candidate_factor_acceptance_replay_metadata,
+)
+from scripts.whitepaper_autoresearch_runner.candidate_board_summaries import (
+    _candidate_board_rejected_signal_outcome_summary as _candidate_board_rejected_signal_outcome_summary,
+    _candidate_spec_requires_order_type_execution_quality as _candidate_spec_requires_order_type_execution_quality,
+    _candidate_spec_requires_predictability_decay_stress as _candidate_spec_requires_predictability_decay_stress,
+    _candidate_board_predictability_decay_summary as _candidate_board_predictability_decay_summary,
+    _candidate_board_scorecard_with_predictability_decay_blockers as _candidate_board_scorecard_with_predictability_decay_blockers,
+    _candidate_board_order_type_execution_quality_summary as _candidate_board_order_type_execution_quality_summary,
+    _candidate_board_scorecard_with_order_type_blockers as _candidate_board_scorecard_with_order_type_blockers,
+    _candidate_spec_requires_queue_position_survival as _candidate_spec_requires_queue_position_survival,
+    _candidate_board_queue_position_survival_summary as _candidate_board_queue_position_survival_summary,
+    _candidate_board_scorecard_with_queue_position_survival_blockers as _candidate_board_scorecard_with_queue_position_survival_blockers,
+    _candidate_board_scorecard_with_rejected_signal_blockers as _candidate_board_scorecard_with_rejected_signal_blockers,
+    _candidate_board_evidence_lineage_summary as _candidate_board_evidence_lineage_summary,
+    _candidate_board_scorecard_with_lineage_blockers as _candidate_board_scorecard_with_lineage_blockers,
+    _candidate_board_replay_window_coverage_summary as _candidate_board_replay_window_coverage_summary,
+    _candidate_board_market_impact_proof_summary as _candidate_board_market_impact_proof_summary,
+    _candidate_board_regime_specialist_summary as _candidate_board_regime_specialist_summary,
+    _candidate_board_scorecard_with_replay_window_blockers as _candidate_board_scorecard_with_replay_window_blockers,
+    _candidate_board_scorecard_with_evidence_blockers as _candidate_board_scorecard_with_evidence_blockers,
+)
+from scripts.whitepaper_autoresearch_runner.candidate_board_status import (
+    _candidate_board_blockers as _candidate_board_blockers,
+    _candidate_board_status as _candidate_board_status,
+    _candidate_board_activity_count as _candidate_board_activity_count,
+    _candidate_board_oracle_blocker_count as _candidate_board_oracle_blocker_count,
+    _candidate_board_net_pnl as _candidate_board_net_pnl,
+    _candidate_board_lower_bound_net_pnl as _candidate_board_lower_bound_net_pnl,
+    _candidate_board_target_progress_ratio as _candidate_board_target_progress_ratio,
+    _candidate_board_required_notional_repair_scale as _candidate_board_required_notional_repair_scale,
+    _candidate_board_best_executed_candidate as _candidate_board_best_executed_candidate,
+    _candidate_board_closest_promotion_candidate as _candidate_board_closest_promotion_candidate,
+    _candidate_board_paper_probation_admission_blockers as _candidate_board_paper_probation_admission_blockers,
+)
+from scripts.whitepaper_autoresearch_runner.common import (
+    _CANDIDATE_BOARD_RUNTIME_SESSION_TZ as _CANDIDATE_BOARD_RUNTIME_SESSION_TZ,
+    _CANDIDATE_BOARD_RUNTIME_SESSION_OPEN as _CANDIDATE_BOARD_RUNTIME_SESSION_OPEN,
+    _CANDIDATE_BOARD_RUNTIME_SESSION_CLOSE as _CANDIDATE_BOARD_RUNTIME_SESSION_CLOSE,
+    _PAPER_PROBATION_LIVE_PAPER_EVIDENCE_REQUIREMENTS as _PAPER_PROBATION_LIVE_PAPER_EVIDENCE_REQUIREMENTS,
+    _PAPER_PROBATION_SAFE_EVIDENCE_COLLECTION_PATH as _PAPER_PROBATION_SAFE_EVIDENCE_COLLECTION_PATH,
+    _SECOND_OOS_WINDOW_ID as _SECOND_OOS_WINDOW_ID,
+    _RUNTIME_CLOSURE_PROOF_ORACLE_BLOCKERS as _RUNTIME_CLOSURE_PROOF_ORACLE_BLOCKERS,
+    _resolve_existing_path as _resolve_existing_path,
+    _stable_hash as _stable_hash,
+    _decimal as _decimal,
+    _decimal_payload as _decimal_payload,
+    _mapping as _mapping,
+    _string as _string,
+    _list_of_mappings as _list_of_mappings,
+    _sequence_of_mappings as _sequence_of_mappings,
+    _rank_sort_value as _rank_sort_value,
+    _proposal_sort_value as _proposal_sort_value,
+    _string_list_from_value as _string_list_from_value,
+    _candidate_board_runtime_ledger_lineage_handoff as _candidate_board_runtime_ledger_lineage_handoff,
+    _candidate_board_runtime_ledger_required_materialized_artifacts as _candidate_board_runtime_ledger_required_materialized_artifacts,
+    _candidate_spec_requires_rejected_signal_outcome_learning as _candidate_spec_requires_rejected_signal_outcome_learning,
+    _boolish as _boolish,
+    _oracle_blockers as _oracle_blockers,
+)
+from scripts.whitepaper_autoresearch_runner.runtime_closure import (
+    _runtime_closure_payload as _runtime_closure_payload,
+    _portfolio_needs_runtime_closure_proof as _portfolio_needs_runtime_closure_proof,
+    _load_json_mapping_artifact as _load_json_mapping_artifact,
+    _runtime_closure_artifact_refs as _runtime_closure_artifact_refs,
+    _runtime_report_summary_int as _runtime_report_summary_int,
+    _runtime_report_int as _runtime_report_int,
+    _runtime_closure_ledger_datetime as _runtime_closure_ledger_datetime,
+    _runtime_closure_exact_replay_bucket_range as _runtime_closure_exact_replay_bucket_range,
+    _runtime_closure_replay_bucket_has_authority as _runtime_closure_replay_bucket_has_authority,
+    _runtime_closure_exact_replay_bucket as _runtime_closure_exact_replay_bucket,
+    _runtime_report_source_markers as _runtime_report_source_markers,
+    _market_impact_default_source_markers as _market_impact_default_source_markers,
+    _runtime_closure_start_equity as _runtime_closure_start_equity,
+    _portfolio_executable_max_notional as _portfolio_executable_max_notional,
+    _runtime_closure_exact_replay_ledger_update as _runtime_closure_exact_replay_ledger_update,
+    _runtime_closure_market_impact_stress_update as _runtime_closure_market_impact_stress_update,
+    _runtime_closure_delay_adjusted_depth_stress_update as _runtime_closure_delay_adjusted_depth_stress_update,
+    _runtime_closure_double_oos_update as _runtime_closure_double_oos_update,
+    _runtime_closure_scorecard_update as _runtime_closure_scorecard_update,
+    _runtime_closure_pending_promotion_steps as _runtime_closure_pending_promotion_steps,
+    _runtime_closure_promotion_prerequisite_blockers as _runtime_closure_promotion_prerequisite_blockers,
+    _promotion_readiness_payload as _promotion_readiness_payload,
+)
+
 
 _DEFAULT_CHIP_UNIVERSE_CSV = ",".join(LIVE_SIGNAL_COVERED_SEMICONDUCTOR_UNIVERSE)
 _DEFAULT_DAILY_PROFIT_TARGET = "500"
@@ -137,9 +252,6 @@ _RANKER_BACKEND_CHOICES = (
 _DEFAULT_PORTFOLIO_PROFIT_PROGRAM = Path(
     "config/trading/research-programs/portfolio-profit-autoresearch-500-v1.yaml"
 )
-_CANDIDATE_BOARD_RUNTIME_SESSION_TZ = ZoneInfo("America/New_York")
-_CANDIDATE_BOARD_RUNTIME_SESSION_OPEN = time(hour=9, minute=30)
-_CANDIDATE_BOARD_RUNTIME_SESSION_CLOSE = time(hour=16, minute=0)
 _DEFAULT_MAX_FRONTIER_CANDIDATES_PER_SPEC = 8
 _DEFAULT_REAL_REPLAY_SHARD_TIMEOUT_SECONDS = 900
 _DEFAULT_REAL_REPLAY_SHARD_WORKERS = 2
@@ -170,23 +282,6 @@ _REJECTED_SIGNAL_OUTCOME_REQUIRED_FIELDS = (
     "post_cost_net_pnl",
     "executable_quote",
 )
-_PAPER_PROBATION_LIVE_PAPER_EVIDENCE_REQUIREMENTS = (
-    "real_runtime_trade_decisions",
-    "broker_order_submissions_and_status_events",
-    "broker_fill_events",
-    "post_cost_costs_tca_and_execution_shortfall",
-    "source_backed_runtime_ledger_lineage",
-    "closed_flat_position_proof",
-    "broker_runtime_ledger_reconciliation",
-)
-_PAPER_PROBATION_SAFE_EVIDENCE_COLLECTION_PATH = (
-    "preserve_final_promotion_gates_fail_closed",
-    "import_exact_replay_runtime_window_metadata_without_live_submit",
-    "run_bounded_live_paper_probe_with_configured_paper_account_only",
-    "materialize_source_backed_runtime_ledger_lineage_for_decisions_orders_fills_costs",
-    "verify_closed_flat_positions_and_broker_runtime_ledger_reconciliation",
-    "reevaluate_post_cost_runtime_profitability_before_any_final_promotion",
-)
 _CODE_COMMIT_ENV_VARS = (
     "TORGHUT_CODE_COMMIT",
     "TORGHUT_COMMIT",
@@ -199,36 +294,6 @@ _CODE_COMMIT_ENV_VARS = (
     "REVISION",
 )
 _PROGRAM_SOURCE_DEFAULT_CONFIDENCE = "0.70"
-_SECOND_OOS_WINDOW_ID = "second_oos"
-_RUNTIME_CLOSURE_PROOF_ORACLE_BLOCKERS = frozenset(
-    {
-        "shadow_parity_status_failed",
-        "executable_replay_passed_failed",
-        "executable_replay_artifact_present_failed",
-        "executable_replay_order_count_failed",
-        "executable_replay_account_buying_power_failed",
-        "executable_replay_max_notional_per_trade_failed",
-        "executable_replay_notional_within_buying_power_failed",
-        "market_impact_stress_passed_failed",
-        "market_impact_stress_artifact_present_failed",
-        "market_impact_liquidity_evidence_present_failed",
-        "market_impact_stress_model_failed",
-        "market_impact_stress_cost_bps_failed",
-        "market_impact_stress_net_pnl_per_day_failed",
-        "delay_adjusted_depth_stress_passed_failed",
-        "delay_adjusted_depth_stress_artifact_present_failed",
-        "delay_adjusted_depth_stress_model_failed",
-        "delay_adjusted_depth_stress_ms_failed",
-        "delay_adjusted_depth_fillable_notional_per_day_failed",
-        "delay_adjusted_depth_stress_net_pnl_per_day_failed",
-        "double_oos_passed_failed",
-        "double_oos_artifact_present_failed",
-        "double_oos_independent_window_count_failed",
-        "double_oos_pass_rate_failed",
-        "double_oos_net_pnl_per_day_failed",
-        "double_oos_cost_shock_net_pnl_per_day_failed",
-    }
-)
 _FAMILY_PRIOR_HARD_BLOCK_ORACLE_BLOCKERS = frozenset(
     {
         "active_day_ratio_below_oracle",
@@ -1664,17 +1729,6 @@ def _dedupe_whitepaper_sources(
     return resolved
 
 
-def _resolve_existing_path(path: Path) -> Path:
-    resolved = path.expanduser().resolve()
-    if path.is_absolute() or resolved.exists():
-        return resolved
-    for parent in Path(__file__).resolve().parents:
-        candidate = (parent / path).resolve()
-        if candidate.exists():
-            return candidate
-    return resolved
-
-
 def _write_failure_summary(
     *,
     output_dir: Path,
@@ -1701,25 +1755,6 @@ def _write_failure_summary(
         summary=summary,
     )
     return summary
-
-
-def _stable_hash(payload: Mapping[str, Any]) -> str:
-    encoded = json.dumps(payload, sort_keys=True, separators=(",", ":"), default=str)
-    return hashlib.sha256(encoded.encode("utf-8")).hexdigest()
-
-
-def _decimal(value: Any, *, default: str = "0") -> Decimal:
-    try:
-        return Decimal(str(value if value is not None else default))
-    except Exception:
-        return Decimal(default)
-
-
-def _decimal_payload(value: Decimal) -> str:
-    text = format(value.normalize(), "f")
-    if "." in text:
-        text = text.rstrip("0").rstrip(".")
-    return text or "0"
 
 
 def _resolved_clickhouse_password(args: argparse.Namespace) -> str:
@@ -1803,18 +1838,6 @@ def _candidate_universe_symbols_for_compilation(
     return symbols
 
 
-def _mapping(value: Any) -> dict[str, Any]:
-    return (
-        {str(key): item for key, item in value.items()}
-        if isinstance(value, Mapping)
-        else {}
-    )
-
-
-def _string(value: Any) -> str:
-    return str(value if value is not None else "").strip()
-
-
 def _current_code_commit() -> str:
     for name in _CODE_COMMIT_ENV_VARS:
         value = _string(os.getenv(name))
@@ -1865,38 +1888,6 @@ def _current_code_commit() -> str:
             dirty = True
             break
     return f"{commit}-dirty" if dirty else commit
-
-
-def _list_of_mappings(value: Any) -> list[Mapping[str, Any]]:
-    if not isinstance(value, list):
-        return []
-    return [
-        cast(Mapping[str, Any], item) for item in value if isinstance(item, Mapping)
-    ]
-
-
-def _sequence_of_mappings(value: Any) -> list[Mapping[str, Any]]:
-    if not isinstance(value, Sequence) or isinstance(value, (str, bytes, bytearray)):
-        return []
-    return [
-        cast(Mapping[str, Any], item)
-        for item in cast(Sequence[Any], value)
-        if isinstance(item, Mapping)
-    ]
-
-
-def _rank_sort_value(value: Any) -> int:
-    try:
-        return int(str(value))
-    except Exception:
-        return 10**9
-
-
-def _proposal_sort_value(value: Any) -> float:
-    try:
-        return float(str(value))
-    except Exception:
-        return 0.0
 
 
 def _scorecard_start_equity(
@@ -3233,13 +3224,6 @@ def _candidate_search_remediation(
     }
 
 
-def _string_list_from_value(value: Any) -> list[str]:
-    if not isinstance(value, Sequence) or isinstance(value, str):
-        return []
-    rows = cast(Sequence[Any], value)
-    return [text for item in rows if (text := _string(item))]
-
-
 def _selected_candidate_spec_ids(
     candidate_selection: Mapping[str, Any],
 ) -> set[str]:
@@ -3314,52 +3298,6 @@ def _candidate_family_goal_rows(
         )
     )
     return payload_rows
-
-
-def _candidate_board_runtime_ledger_lineage_handoff(
-    *,
-    scorecard: Mapping[str, Any],
-    evidence: CandidateEvidenceBundle | None,
-) -> dict[str, Any]:
-    scorecard_handoff = _mapping(
-        scorecard.get("runtime_ledger_lineage_materialization_handoff")
-    )
-    if scorecard_handoff:
-        return scorecard_handoff
-    if evidence is None:
-        return {}
-    promotion_readiness = _mapping(evidence.promotion_readiness)
-    return _mapping(
-        promotion_readiness.get("runtime_ledger_lineage_materialization_handoff")
-    )
-
-
-def _candidate_board_runtime_ledger_required_materialized_artifacts(
-    handoff: Mapping[str, Any],
-) -> list[str]:
-    raw_artifacts = handoff.get("required_materialized_artifacts") or handoff.get(
-        "required_artifacts"
-    )
-    if isinstance(raw_artifacts, str):
-        artifacts = [_string(raw_artifacts)]
-    elif isinstance(raw_artifacts, Sequence):
-        artifacts = []
-        for item in cast(Sequence[Any], raw_artifacts):
-            if isinstance(item, Mapping):
-                artifact = _string(
-                    item.get("artifact_ref")
-                    or item.get("ref")
-                    or item.get("path")
-                    or item.get("name")
-                    or item.get("artifact")
-                )
-            else:
-                artifact = _string(item)
-            if artifact:
-                artifacts.append(artifact)
-    else:
-        artifacts = []
-    return list(dict.fromkeys(artifacts))
 
 
 def _candidate_sleeve_goal_proof_handoff_fields(
@@ -4282,20 +4220,6 @@ def _candidate_spec_is_false_negative_rescue(spec: CandidateSpec) -> bool:
         and _string(params.get("veto_relaxation_scope"))
         == "labeled_false_negative_only"
         and _string(params.get("outcome_label_filter")) == "profitable_after_costs"
-    )
-
-
-def _candidate_spec_requires_rejected_signal_outcome_learning(
-    spec: CandidateSpec,
-) -> bool:
-    overlays = spec.parameter_space.get("mechanism_overlay_ids", ())
-    overlay_ids = set(_string_list_from_value(overlays))
-    return (
-        "rejected_signal_outcome_calibration" in overlay_ids
-        or _boolish(
-            spec.promotion_contract.get("requires_rejected_signal_outcome_learning")
-        )
-        or "required_min_rejected_signal_outcome_label_count" in spec.hard_vetoes
     )
 
 
@@ -8749,142 +8673,6 @@ def _epoch_mlx_snapshot_manifest(
     )
 
 
-def _runtime_closure_payload(
-    *,
-    args: argparse.Namespace,
-    output_dir: Path,
-    epoch_id: str,
-    program: StrategyAutoresearchProgram,
-    manifest: MlxSnapshotManifest,
-    portfolio: PortfolioCandidateSpec | None,
-) -> Mapping[str, Any]:
-    execution_context = RuntimeClosureExecutionContext(
-        strategy_configmap_path=_resolve_existing_path(args.strategy_configmap),
-        clickhouse_http_url=str(args.clickhouse_http_url),
-        clickhouse_username=str(args.clickhouse_username)
-        if args.clickhouse_username
-        else None,
-        clickhouse_password=str(args.clickhouse_password)
-        if args.clickhouse_password
-        else None,
-        start_equity=_decimal(args.start_equity, default="31590.02"),
-        chunk_minutes=int(args.chunk_minutes),
-        symbols=tuple(
-            symbol.strip() for symbol in str(args.symbols).split(",") if symbol.strip()
-        ),
-        progress_log_interval_seconds=int(args.progress_log_seconds),
-        shadow_validation_artifact_path=_resolve_existing_path(
-            cast(Path, getattr(args, "shadow_validation_artifact"))
-        )
-        if getattr(args, "shadow_validation_artifact", None) is not None
-        else None,
-    )
-    return write_runtime_closure_bundle(
-        run_root=output_dir,
-        runner_run_id=epoch_id,
-        program=program,
-        best_candidate=portfolio,
-        manifest=manifest,
-        execution_context=execution_context if portfolio is not None else None,
-    ).to_payload()
-
-
-def _boolish(value: Any) -> bool:
-    if isinstance(value, bool):
-        return value
-    return _string(value).lower() in {"1", "true", "yes", "y", "passed"}
-
-
-def _oracle_blockers(scorecard: Mapping[str, Any]) -> frozenset[str]:
-    raw_blockers = _mapping(scorecard.get("profit_target_oracle")).get("blockers")
-    if not isinstance(raw_blockers, Sequence) or isinstance(raw_blockers, str):
-        return frozenset()
-    return frozenset(
-        blocker
-        for blocker in (_string(item) for item in cast(Sequence[Any], raw_blockers))
-        if blocker
-    )
-
-
-def _portfolio_needs_runtime_closure_proof(portfolio: PortfolioCandidateSpec) -> bool:
-    scorecard = _mapping(portfolio.objective_scorecard)
-    if _boolish(scorecard.get("oracle_passed")):
-        return False
-    if not _boolish(scorecard.get("target_met")):
-        return False
-    blockers = _oracle_blockers(scorecard)
-    return bool(blockers) and blockers <= _RUNTIME_CLOSURE_PROOF_ORACLE_BLOCKERS
-
-
-def _load_json_mapping_artifact(path_value: Any) -> dict[str, Any]:
-    path_text = _string(path_value)
-    if not path_text:
-        return {}
-    try:
-        payload = json.loads(Path(path_text).read_text(encoding="utf-8"))
-    except (FileNotFoundError, json.JSONDecodeError, OSError):
-        return {}
-    return _mapping(payload)
-
-
-def _runtime_closure_artifact_refs(
-    runtime_closure: Mapping[str, Any],
-) -> tuple[str, ...]:
-    refs: list[str] = []
-    root = _string(runtime_closure.get("root"))
-    if root:
-        summary_path = Path(root) / "summary.json"
-        if summary_path.exists():
-            refs.append(str(summary_path))
-    for key in (
-        "candidate_configmap_path",
-        "gate_report_path",
-        "parity_replay_path",
-        "parity_report_path",
-        "approval_replay_path",
-        "approval_report_path",
-        "exact_replay_ledger_artifact_path",
-        "exact_replay_ledger_artifact_ref",
-        "runtime_ledger_artifact_path",
-        "runtime_ledger_artifact_ref",
-        "shadow_validation_path",
-        "portfolio_proof_receipt_path",
-        "market_impact_stress_report_path",
-        "market_impact_stress_artifact_path",
-        "delay_adjusted_depth_stress_report_path",
-        "delay_depth_stress_report_path",
-        "latency_depth_stress_artifact_path",
-        "double_oos_report_path",
-        "double_oos_artifact_path",
-        "walkforward_double_oos_report_path",
-        "stress_metrics_path",
-        "profitability_stage_manifest_path",
-        "promotion_prerequisites_path",
-        "replay_plan_path",
-    ):
-        ref = _string(runtime_closure.get(key))
-        if ref and Path(ref).exists() and ref not in refs:
-            refs.append(ref)
-    return tuple(refs)
-
-
-def _runtime_report_summary_int(
-    report: Mapping[str, Any], key: str, *, default: int = 0
-) -> int:
-    value = _mapping(report.get("summary")).get(key)
-    try:
-        return max(0, int(Decimal(str(value if value is not None else default))))
-    except Exception:
-        return default
-
-
-def _runtime_report_int(value: Any, *, default: int = 0) -> int:
-    try:
-        return max(0, int(Decimal(str(value if value is not None else default))))
-    except Exception:
-        return default
-
-
 _EXACT_REPLAY_LEDGER_ARTIFACT_KIND = "exact_replay_ledger"
 _EXACT_REPLAY_LEDGER_SCHEMA_VERSIONS = frozenset(
     {
@@ -8893,3153 +8681,6 @@ _EXACT_REPLAY_LEDGER_SCHEMA_VERSIONS = frozenset(
     }
 )
 _EXACT_REPLAY_RUNTIME_LEDGER_PNL_SOURCE = "exact_replay_runtime_ledger"
-
-
-def _runtime_closure_ledger_datetime(
-    value: Any, *, date_end: bool = False
-) -> datetime | None:
-    text = _string(value)
-    if not text:
-        return None
-    try:
-        if "T" not in text and len(text) == 10:
-            parsed_date = date.fromisoformat(text)
-            parsed = datetime.combine(parsed_date, time.min, tzinfo=UTC)
-            return parsed + timedelta(days=1) if date_end else parsed
-        parsed = datetime.fromisoformat(text.replace("Z", "+00:00"))
-    except ValueError:
-        return None
-    if parsed.tzinfo is None:
-        return parsed.replace(tzinfo=UTC)
-    return parsed.astimezone(UTC)
-
-
-def _runtime_closure_exact_replay_bucket_range(
-    ledger: Mapping[str, Any],
-) -> tuple[datetime, datetime] | None:
-    start = _runtime_closure_ledger_datetime(
-        ledger.get("window_start")
-        or ledger.get("bucket_started_at")
-        or ledger.get("started_at")
-        or ledger.get("start"),
-    )
-    end = _runtime_closure_ledger_datetime(
-        ledger.get("window_end")
-        or ledger.get("bucket_ended_at")
-        or ledger.get("ended_at")
-        or ledger.get("end"),
-        date_end=True,
-    )
-    if start is None or end is None or end <= start:
-        return None
-    return (start, end)
-
-
-def _runtime_closure_replay_bucket_has_authority(bucket: RuntimeLedgerBucket) -> bool:
-    return (
-        not bucket.blockers
-        and bucket.fill_count > 0
-        and bucket.decision_count > 0
-        and bucket.submitted_order_count > 0
-        and bucket.closed_trade_count > 0
-        and bucket.open_position_count == 0
-        and bucket.filled_notional > 0
-        and bool(bucket.cost_basis_counts)
-        and bool(bucket.execution_policy_hash_counts)
-        and bool(bucket.cost_model_hash_counts)
-        and bool(bucket.lineage_hash_counts)
-        and bucket.pnl_basis == POST_COST_PNL_BASIS
-    )
-
-
-def _runtime_closure_exact_replay_bucket(
-    *,
-    ledger: Mapping[str, Any],
-    rows: Sequence[Mapping[str, object]],
-) -> RuntimeLedgerBucket | None:
-    bucket_range = _runtime_closure_exact_replay_bucket_range(ledger)
-    if bucket_range is None:
-        return None
-    buckets = build_runtime_ledger_buckets(
-        rows,
-        bucket_ranges=[bucket_range],
-        require_order_lifecycle=True,
-    )
-    if len(buckets) != 1:
-        return None
-    bucket = buckets[0]
-    if not _runtime_closure_replay_bucket_has_authority(bucket):
-        return None
-    return bucket
-
-
-def _runtime_report_source_markers(report: Mapping[str, Any]) -> list[str]:
-    return sorted(set(_string_list_from_value(report.get("source_markers"))))
-
-
-def _market_impact_default_source_markers() -> list[str]:
-    return [
-        "double_square_root_impact_arxiv_2502_16246_2025",
-        "realistic_market_impact_arxiv_2603_29086_2026",
-    ]
-
-
-def _runtime_closure_start_equity(runtime_closure: Mapping[str, Any]) -> Decimal:
-    replay_plan = _load_json_mapping_artifact(runtime_closure.get("replay_plan_path"))
-    execution_context = _mapping(replay_plan.get("execution_context"))
-    return _decimal(execution_context.get("start_equity"))
-
-
-def _portfolio_executable_max_notional(portfolio: PortfolioCandidateSpec) -> Decimal:
-    max_notional = Decimal("0")
-    base_per_leg_notional = Decimal("50000")
-    for sleeve in portfolio.sleeves:
-        explicit_max = _decimal(sleeve.get("max_notional_per_trade"))
-        if explicit_max > 0:
-            max_notional = max(max_notional, explicit_max)
-            continue
-        weight = _decimal(sleeve.get("weight"), default="1")
-        if weight <= 0:
-            weight = Decimal("1")
-        max_notional = max(max_notional, base_per_leg_notional * weight)
-    return max_notional
-
-
-def _runtime_closure_exact_replay_ledger_update(
-    runtime_closure: Mapping[str, Any],
-) -> dict[str, Any]:
-    # Exact-replay rows are useful diagnostics, but they are not runtime or
-    # live-paper ledger authority. Final proof must come from source-backed
-    # runtime-ledger materialization instead of replay artifacts.
-    return {}
-
-
-def _runtime_closure_market_impact_stress_update(
-    runtime_closure: Mapping[str, Any],
-) -> dict[str, Any]:
-    market_impact_report_path = _string(
-        runtime_closure.get("market_impact_stress_report_path")
-        or runtime_closure.get("market_impact_stress_artifact_path")
-    )
-    market_impact_report = _load_json_mapping_artifact(market_impact_report_path)
-    if not market_impact_report:
-        return {}
-    model = _string(
-        market_impact_report.get("model")
-        or market_impact_report.get("impact_model")
-        or market_impact_report.get("cost_model")
-    )
-    cost_bps = str(
-        _decimal(
-            market_impact_report.get("impact_cost_bps")
-            or market_impact_report.get("market_impact_cost_bps")
-            or market_impact_report.get("cost_shock_bps")
-        )
-    )
-    net_pnl_per_day = str(
-        _decimal(
-            market_impact_report.get("post_impact_net_pnl_per_day")
-            or market_impact_report.get("stressed_net_pnl_per_day")
-            or market_impact_report.get("net_pnl_per_day")
-        )
-    )
-    components = _mapping(market_impact_report.get("market_impact_stress_components"))
-    source_markers = _runtime_report_source_markers(market_impact_report)
-    if not source_markers:
-        source_markers = sorted(
-            set(_string_list_from_value(components.get("source_markers")))
-        )
-    if not source_markers and (components or model):
-        source_markers = _market_impact_default_source_markers()
-    if not components and model and _decimal(cost_bps) > 0:
-        components = {
-            "selected_model": model,
-            "selected_cost_bps": cost_bps,
-            "source_marker": "realistic_market_impact_arxiv_2603_29086_2026",
-            "source_markers": source_markers,
-        }
-    elif components and source_markers and "source_markers" not in components:
-        components = {**components, "source_markers": source_markers}
-    return {
-        "market_impact_stress_passed": _boolish(
-            market_impact_report.get("objective_met")
-            or market_impact_report.get("passed")
-        ),
-        "market_impact_stress_artifact_ref": market_impact_report_path,
-        "market_impact_stress_source_markers": source_markers,
-        "market_impact_stress_model": model,
-        "market_impact_stress_cost_bps": cost_bps,
-        "market_impact_stress_components": components,
-        "nonlinear_market_impact_stress_passed": _boolish(
-            market_impact_report.get("objective_met")
-            or market_impact_report.get("passed")
-        ),
-        "nonlinear_market_impact_stress_model": model,
-        "nonlinear_market_impact_stress_cost_bps": cost_bps,
-        "nonlinear_market_impact_stress_net_pnl_per_day": net_pnl_per_day,
-        "permanent_impact_decay_model": _string(
-            market_impact_report.get("permanent_impact_decay_model")
-            or "exponential_decay_proxy"
-        ),
-        "market_impact_liquidity_evidence_present": _boolish(
-            market_impact_report.get("liquidity_evidence_present")
-            or market_impact_report.get("market_impact_liquidity_evidence_present")
-        ),
-        "market_impact_stress_net_pnl_per_day": net_pnl_per_day,
-    }
-
-
-def _runtime_closure_delay_adjusted_depth_stress_update(
-    runtime_closure: Mapping[str, Any],
-) -> dict[str, Any]:
-    delay_depth_report_path = _string(
-        runtime_closure.get("delay_adjusted_depth_stress_report_path")
-        or runtime_closure.get("delay_depth_stress_report_path")
-        or runtime_closure.get("latency_depth_stress_artifact_path")
-    )
-    delay_depth_report = _load_json_mapping_artifact(delay_depth_report_path)
-    if not delay_depth_report:
-        return {}
-    checked_at = _string(
-        delay_depth_report.get("generated_at") or delay_depth_report.get("checked_at")
-    )
-    fillable_notional_per_day = _decimal(
-        delay_depth_report.get("fillable_notional_per_day")
-        or delay_depth_report.get("depth_fillable_notional_per_day")
-    )
-    stress_ms = _decimal(
-        delay_depth_report.get("stress_delay_ms") or delay_depth_report.get("delay_ms")
-    )
-    latency_grid = cast(
-        Sequence[Any],
-        delay_depth_report.get("latency_grid_ms")
-        or delay_depth_report.get("delay_grid_ms")
-        or (),
-    )
-    if not latency_grid and stress_ms > 0:
-        latency_grid = ("50", "150", "250")
-    fill_survival_sample_count = max(
-        _runtime_report_int(
-            delay_depth_report.get("delay_adjusted_depth_fill_survival_sample_count")
-        ),
-        _runtime_report_int(delay_depth_report.get("fill_survival_sample_count")),
-    )
-    fill_survival_rate = _decimal(
-        delay_depth_report.get("delay_adjusted_depth_fill_survival_rate")
-        or delay_depth_report.get("fill_survival_fill_rate")
-        or delay_depth_report.get("fill_survival_rate")
-    )
-    fill_survival_evidence_present = _boolish(
-        delay_depth_report.get("delay_adjusted_depth_fill_survival_evidence_present")
-        or delay_depth_report.get("fill_survival_evidence_present")
-        or (fill_survival_sample_count > 0)
-    )
-    queue_position_survival_evidence_present = _boolish(
-        delay_depth_report.get("queue_position_survival_fill_curve_evidence_present")
-        or delay_depth_report.get("queue_position_survival_evidence_present")
-    )
-    queue_position_survival_sample_count = max(
-        _runtime_report_int(
-            delay_depth_report.get("queue_position_survival_sample_count")
-        ),
-        _runtime_report_int(
-            delay_depth_report.get("queue_position_survival_fill_sample_count")
-        ),
-    )
-    if (
-        queue_position_survival_evidence_present
-        and queue_position_survival_sample_count == 0
-    ):
-        queue_position_survival_sample_count = fill_survival_sample_count
-    queue_position_survival_fill_rate = _decimal(
-        delay_depth_report.get("queue_position_survival_fill_rate")
-        or (
-            fill_survival_rate
-            if queue_position_survival_evidence_present
-            else Decimal("0")
-        )
-    )
-    queue_position_survival_queue_ratio_p95 = _decimal(
-        delay_depth_report.get("queue_position_survival_queue_ratio_p95")
-    )
-    queue_ahead_depletion_sample_count = max(
-        _runtime_report_int(
-            delay_depth_report.get(
-                "queue_position_survival_queue_ahead_depletion_sample_count"
-            )
-        ),
-        _runtime_report_int(
-            delay_depth_report.get(
-                "delay_adjusted_depth_queue_ahead_depletion_sample_count"
-            )
-        ),
-        _runtime_report_int(
-            delay_depth_report.get("queue_ahead_depletion_sample_count")
-        ),
-    )
-    queue_ahead_depletion_evidence_present = _boolish(
-        delay_depth_report.get(
-            "queue_position_survival_queue_ahead_depletion_evidence_present"
-        )
-        or delay_depth_report.get(
-            "delay_adjusted_depth_queue_ahead_depletion_evidence_present"
-        )
-        or delay_depth_report.get("queue_ahead_depletion_evidence_present")
-        or (queue_ahead_depletion_sample_count > 0)
-    )
-    queue_position_survival_net_pnl_per_day = _decimal(
-        delay_depth_report.get(
-            "post_cost_net_pnl_after_queue_position_survival_fill_stress"
-        )
-        or delay_depth_report.get("post_queue_position_survival_net_pnl_per_day")
-        or (
-            delay_depth_report.get("post_delay_depth_net_pnl_per_day")
-            if queue_position_survival_evidence_present
-            and queue_ahead_depletion_evidence_present
-            else Decimal("0")
-        )
-    )
-    return {
-        "delay_adjusted_depth_stress_checks_total": max(
-            _runtime_report_int(delay_depth_report.get("stress_case_count")),
-            _runtime_report_int(delay_depth_report.get("case_count")),
-            _runtime_report_int(delay_depth_report.get("trading_day_count")),
-        ),
-        "delay_adjusted_depth_stress_passed": _boolish(
-            delay_depth_report.get("objective_met") or delay_depth_report.get("passed")
-        ),
-        "delay_adjusted_depth_stress_checked_at": checked_at,
-        "delay_adjusted_depth_stress_artifact_ref": delay_depth_report_path,
-        "delay_adjusted_depth_stress_source_markers": _runtime_report_source_markers(
-            delay_depth_report
-        ),
-        "delay_adjusted_depth_stress_report": dict(delay_depth_report),
-        "delay_adjusted_depth_stress_model": _string(
-            delay_depth_report.get("model") or delay_depth_report.get("stress_model")
-        ),
-        "delay_adjusted_depth_stress_ms": str(stress_ms),
-        "delay_adjusted_depth_latency_grid_ms": [str(item) for item in latency_grid],
-        "delay_adjusted_depth_grid_max_stress_ms": str(
-            max((_decimal(item) for item in latency_grid), default=stress_ms)
-        ),
-        "delay_adjusted_depth_liquidity_evidence_present": bool(
-            fillable_notional_per_day > 0
-        ),
-        "delay_adjusted_depth_liquidity_missing_day_count": _runtime_report_int(
-            delay_depth_report.get("missing_liquidity_day_count")
-        ),
-        "delay_adjusted_depth_fillable_notional_per_day": str(
-            fillable_notional_per_day
-        ),
-        "delay_adjusted_depth_worst_active_day_fillable_notional": str(
-            _decimal(
-                delay_depth_report.get("worst_active_day_fillable_notional")
-                or fillable_notional_per_day
-            )
-        ),
-        "delay_adjusted_depth_p10_active_day_fillable_notional": str(
-            _decimal(
-                delay_depth_report.get("p10_active_day_fillable_notional")
-                or fillable_notional_per_day
-            )
-        ),
-        "delay_adjusted_depth_tail_coverage_passed": _boolish(
-            delay_depth_report.get("tail_coverage_passed")
-            if "tail_coverage_passed" in delay_depth_report
-            else fillable_notional_per_day > 0
-        ),
-        "delay_adjusted_depth_fill_survival_evidence_present": fill_survival_evidence_present,
-        "delay_adjusted_depth_fill_survival_sample_count": fill_survival_sample_count,
-        "delay_adjusted_depth_fill_survival_rate": str(fill_survival_rate),
-        "queue_position_survival_fill_curve_evidence_present": (
-            queue_position_survival_evidence_present
-        ),
-        "queue_position_survival_sample_count": queue_position_survival_sample_count,
-        "queue_position_survival_fill_rate": str(queue_position_survival_fill_rate),
-        "queue_position_survival_queue_ratio_p95": str(
-            queue_position_survival_queue_ratio_p95
-        ),
-        "queue_position_survival_queue_ahead_depletion_evidence_present": (
-            queue_ahead_depletion_evidence_present
-        ),
-        "queue_position_survival_queue_ahead_depletion_sample_count": (
-            queue_ahead_depletion_sample_count
-        ),
-        "delay_adjusted_depth_queue_ahead_depletion_evidence_present": (
-            queue_ahead_depletion_evidence_present
-        ),
-        "delay_adjusted_depth_queue_ahead_depletion_sample_count": (
-            queue_ahead_depletion_sample_count
-        ),
-        "queue_ahead_depletion_evidence_present": queue_ahead_depletion_evidence_present,
-        "queue_ahead_depletion_sample_count": queue_ahead_depletion_sample_count,
-        "delay_adjusted_depth_stress_net_pnl_per_day": str(
-            _decimal(
-                delay_depth_report.get("post_delay_depth_net_pnl_per_day")
-                or delay_depth_report.get("stressed_net_pnl_per_day")
-            )
-        ),
-        "post_cost_net_pnl_after_queue_position_survival_fill_stress": str(
-            queue_position_survival_net_pnl_per_day
-        ),
-    }
-
-
-def _runtime_closure_double_oos_update(
-    runtime_closure: Mapping[str, Any],
-) -> dict[str, Any]:
-    double_oos_report_path = _string(
-        runtime_closure.get("double_oos_report_path")
-        or runtime_closure.get("double_oos_artifact_path")
-        or runtime_closure.get("walkforward_double_oos_report_path")
-    )
-    double_oos_report = _load_json_mapping_artifact(double_oos_report_path)
-    if not double_oos_report:
-        return {}
-    return {
-        "double_oos_passed": _boolish(
-            double_oos_report.get("objective_met") or double_oos_report.get("passed")
-        ),
-        "double_oos_artifact_ref": double_oos_report_path,
-        "double_oos_source_markers": _runtime_report_source_markers(double_oos_report),
-        "double_oos_independent_window_count": max(
-            _runtime_report_int(double_oos_report.get("independent_window_count")),
-            _runtime_report_int(double_oos_report.get("window_count")),
-            _runtime_report_int(double_oos_report.get("fold_count")),
-        ),
-        "double_oos_pass_rate": str(_decimal(double_oos_report.get("pass_rate"))),
-        "double_oos_net_pnl_per_day": str(
-            _decimal(
-                double_oos_report.get("net_pnl_per_day")
-                or double_oos_report.get("post_double_oos_net_pnl_per_day")
-            )
-        ),
-        "double_oos_cost_shock_net_pnl_per_day": str(
-            _decimal(
-                double_oos_report.get("cost_shock_net_pnl_per_day")
-                or double_oos_report.get("post_cost_shock_net_pnl_per_day")
-                or double_oos_report.get("market_impact_stress_net_pnl_per_day")
-            )
-        ),
-    }
-
-
-def _runtime_closure_scorecard_update(
-    *,
-    portfolio: PortfolioCandidateSpec,
-    runtime_closure: Mapping[str, Any],
-) -> dict[str, Any]:
-    parity_report = _load_json_mapping_artifact(
-        runtime_closure.get("parity_report_path")
-    )
-    approval_report = _load_json_mapping_artifact(
-        runtime_closure.get("approval_report_path")
-    )
-    shadow_validation = _load_json_mapping_artifact(
-        runtime_closure.get("shadow_validation_path")
-    )
-    parity_pass = _boolish(parity_report.get("objective_met"))
-    approval_pass = _boolish(approval_report.get("objective_met"))
-    artifact_refs = _runtime_closure_artifact_refs(runtime_closure)
-    executable_artifact_refs = tuple(
-        ref
-        for ref in (
-            _string(runtime_closure.get("parity_replay_path")),
-            _string(runtime_closure.get("approval_replay_path")),
-        )
-        if ref and Path(ref).exists()
-    )
-    executable_report = approval_report if approval_report else parity_report
-    filled_order_count = _runtime_report_summary_int(executable_report, "filled_count")
-    submitted_order_count = _runtime_report_summary_int(
-        executable_report, "decision_count"
-    )
-    shadow_status = _string(shadow_validation.get("status")) or "missing"
-    market_impact_update = _runtime_closure_market_impact_stress_update(runtime_closure)
-    delay_depth_update = _runtime_closure_delay_adjusted_depth_stress_update(
-        runtime_closure
-    )
-    double_oos_update = _runtime_closure_double_oos_update(runtime_closure)
-    exact_replay_ledger_update = _runtime_closure_exact_replay_ledger_update(
-        runtime_closure
-    )
-    runtime_closure_source_markers = sorted(
-        set(
-            _string_list_from_value(
-                market_impact_update.get("market_impact_stress_source_markers")
-            )
-            + _string_list_from_value(
-                delay_depth_update.get("delay_adjusted_depth_stress_source_markers")
-            )
-            + _string_list_from_value(
-                double_oos_update.get("double_oos_source_markers")
-            )
-        )
-    )
-    return {
-        "runtime_closure_proof": {
-            "status": _string(runtime_closure.get("status")) or "missing",
-            "parity_objective_met": parity_pass,
-            "approval_objective_met": approval_pass,
-            "shadow_status": shadow_status,
-            "artifact_refs": list(artifact_refs),
-            "executable_replay_artifact_refs": list(executable_artifact_refs),
-            "market_impact_stress_artifact_ref": market_impact_update.get(
-                "market_impact_stress_artifact_ref", ""
-            ),
-            "delay_adjusted_depth_stress_artifact_ref": delay_depth_update.get(
-                "delay_adjusted_depth_stress_artifact_ref", ""
-            ),
-            "double_oos_artifact_ref": double_oos_update.get(
-                "double_oos_artifact_ref", ""
-            ),
-            "exact_replay_ledger_artifact_ref": exact_replay_ledger_update.get(
-                "exact_replay_ledger_artifact_ref", ""
-            ),
-        },
-        "runtime_closure_artifact_refs": list(artifact_refs),
-        "shadow_parity_status": "within_budget"
-        if shadow_status == "within_budget"
-        else shadow_status,
-        "executable_replay_passed": (
-            parity_pass and approval_pass and bool(executable_artifact_refs)
-        ),
-        "executable_replay_artifact_refs": list(executable_artifact_refs),
-        "executable_replay_artifact_ref": executable_artifact_refs[-1]
-        if executable_artifact_refs
-        else "",
-        "executable_replay_order_count": filled_order_count,
-        "executable_replay_submitted_order_count": submitted_order_count,
-        "executable_replay_account_buying_power": str(
-            _runtime_closure_start_equity(runtime_closure)
-        ),
-        "executable_replay_max_notional_per_trade": str(
-            _portfolio_executable_max_notional(portfolio)
-        ),
-        "runtime_closure_source_markers": runtime_closure_source_markers,
-        **exact_replay_ledger_update,
-        **market_impact_update,
-        **delay_depth_update,
-        **double_oos_update,
-    }
-
-
-def _runtime_closure_pending_promotion_steps(
-    runtime_closure: Mapping[str, Any],
-) -> tuple[str, ...]:
-    return tuple(
-        step
-        for step in (
-            _string(item)
-            for item in cast(
-                Sequence[Any], runtime_closure.get("next_required_steps") or ()
-            )
-        )
-        if step and step != "promotion_review"
-    )
-
-
-def _runtime_closure_promotion_prerequisite_blockers(
-    runtime_closure: Mapping[str, Any],
-) -> tuple[str, ...]:
-    prerequisites = _mapping(runtime_closure.get("promotion_prerequisites"))
-    if not prerequisites:
-        return ("promotion_prerequisites_missing",)
-    if _boolish(prerequisites.get("allowed")):
-        return ()
-    reasons = tuple(
-        reason
-        for reason in (
-            _string(item)
-            for item in cast(Sequence[Any], prerequisites.get("reasons") or ())
-        )
-        if reason
-    )
-    return reasons or ("promotion_prerequisites_denied",)
-
-
-def _promotion_readiness_payload(
-    *,
-    oracle_candidate_found: bool,
-    status: str,
-    blockers: Sequence[str],
-    runtime_closure: Mapping[str, Any],
-) -> dict[str, Any]:
-    blocker_list = list(
-        dict.fromkeys(_string(item) for item in blockers if _string(item))
-    )
-    if oracle_candidate_found:
-        blocker_list = list(
-            dict.fromkeys(
-                (
-                    *blocker_list,
-                    *_runtime_closure_pending_promotion_steps(runtime_closure),
-                    *_runtime_closure_promotion_prerequisite_blockers(runtime_closure),
-                )
-            )
-        )
-    if oracle_candidate_found and not blocker_list:
-        return {
-            "status": "promotion_ready",
-            "promotable": True,
-            "blockers": [],
-        }
-    return {
-        "status": "blocked_pending_promotion_prerequisites"
-        if oracle_candidate_found and blocker_list
-        else status,
-        "promotable": False,
-        "blockers": blocker_list,
-    }
-
-
-def _candidate_board_score_rows(
-    rows: Sequence[Mapping[str, Any]],
-) -> dict[str, Mapping[str, Any]]:
-    return {
-        _string(row.get("candidate_spec_id")): row
-        for row in rows
-        if _string(row.get("candidate_spec_id"))
-    }
-
-
-def _candidate_board_decimal_field(scorecard: Mapping[str, Any], key: str) -> str:
-    value = scorecard.get(key)
-    if value is None:
-        return ""
-    return str(value)
-
-
-def _candidate_board_int_field(scorecard: Mapping[str, Any], key: str) -> int:
-    try:
-        return max(0, int(Decimal(str(scorecard.get(key, 0) or 0))))
-    except Exception:
-        return 0
-
-
-def _candidate_board_first_int_field(
-    scorecard: Mapping[str, Any], keys: Sequence[str]
-) -> int:
-    for key in keys:
-        value = _candidate_board_int_field(scorecard, key)
-        if value > 0:
-            return value
-    return 0
-
-
-def _candidate_board_rejected_signal_outcome_summary(
-    spec: CandidateSpec, scorecard: Mapping[str, Any]
-) -> dict[str, Any]:
-    required = _candidate_spec_requires_rejected_signal_outcome_learning(spec)
-    required_fields = tuple(
-        _string_list_from_value(
-            spec.hard_vetoes.get("required_rejected_signal_counterfactual_fields")
-        )
-        or (
-            "counterfactual_return",
-            "route_tca",
-            "post_cost_net_pnl",
-            "executable_quote",
-        )
-    )
-    observed_fields = set(
-        _string_list_from_value(
-            scorecard.get("rejected_signal_counterfactual_fields")
-            or scorecard.get("rejected_signal_outcome_counterfactual_fields")
-            or scorecard.get("rejected_signal_counterfactual_fields_present")
-        )
-    )
-    if _boolish(scorecard.get("rejected_signal_counterfactual_fields_present")):
-        observed_fields.update(required_fields)
-    min_label_count = _candidate_board_int_field(
-        spec.hard_vetoes, "required_min_rejected_signal_outcome_label_count"
-    )
-    if min_label_count <= 0:
-        min_label_count = 120
-    max_pending_ratio = _decimal(
-        spec.hard_vetoes.get("required_max_rejected_signal_outcome_pending_ratio"),
-        default="0.05",
-    )
-    min_reason_coverage = _decimal(
-        spec.hard_vetoes.get("required_min_rejected_signal_reason_coverage"),
-        default="0.80",
-    )
-    required_persistence_state = _string(
-        spec.hard_vetoes.get("required_rejected_signal_outcome_persistence_state")
-        or "ok"
-    ).lower()
-    label_count = _candidate_board_first_int_field(
-        scorecard,
-        (
-            "rejected_signal_outcome_labeled_count",
-            "rejected_signal_outcome_label_count",
-            "rejected_signal_outcome_labeled_event_count",
-        ),
-    )
-    pending_ratio = _decimal(
-        scorecard.get("rejected_signal_outcome_pending_ratio"), default="1"
-    )
-    reason_coverage = _decimal(
-        scorecard.get("rejected_signal_reason_coverage")
-        or scorecard.get("rejected_signal_outcome_reason_coverage")
-    )
-    persistence_state = _string(
-        scorecard.get("rejected_signal_outcome_persistence_state")
-        or scorecard.get("rejected_signal_persistence_state")
-    ).lower()
-    blockers: list[str] = []
-    if required:
-        if label_count < min_label_count:
-            blockers.append("rejected_signal_outcome_labeled_count_failed")
-        if pending_ratio > max_pending_ratio:
-            blockers.append("rejected_signal_outcome_pending_ratio_failed")
-        if reason_coverage < min_reason_coverage:
-            blockers.append("rejected_signal_reason_coverage_failed")
-        if not set(required_fields).issubset(observed_fields):
-            blockers.append("rejected_signal_counterfactual_fields_present_failed")
-        if persistence_state != required_persistence_state:
-            blockers.append("rejected_signal_outcome_persistence_state_failed")
-    return {
-        "required": required,
-        "passed": not blockers,
-        "blockers": blockers,
-        "labeled_count": label_count,
-        "min_labeled_count": min_label_count,
-        "pending_ratio": str(pending_ratio),
-        "max_pending_ratio": str(max_pending_ratio),
-        "reason_coverage": str(reason_coverage),
-        "min_reason_coverage": str(min_reason_coverage),
-        "persistence_state": persistence_state,
-        "required_persistence_state": required_persistence_state,
-        "counterfactual_fields": sorted(observed_fields),
-        "required_counterfactual_fields": list(required_fields),
-    }
-
-
-def _candidate_spec_requires_order_type_execution_quality(spec: CandidateSpec) -> bool:
-    overlay_ids = set(
-        _string_list_from_value(spec.parameter_space.get("mechanism_overlay_ids"))
-    )
-    return (
-        "mixed_market_limit_execution_policy" in overlay_ids
-        or _boolish(
-            spec.promotion_contract.get("requires_order_type_execution_quality")
-        )
-        or _boolish(spec.promotion_contract.get("requires_order_type_ablation"))
-        or _boolish(spec.promotion_contract.get("requires_market_limit_order_mix"))
-        or "required_order_type_ablation_passed" in spec.hard_vetoes
-        or "required_market_limit_order_mix_evidence" in spec.hard_vetoes
-    )
-
-
-def _candidate_spec_requires_predictability_decay_stress(
-    spec: CandidateSpec,
-) -> bool:
-    overlay_ids = set(
-        _string_list_from_value(spec.parameter_space.get("mechanism_overlay_ids"))
-    )
-    return (
-        "alpha_decay_predictability_stress" in overlay_ids
-        or _boolish(spec.promotion_contract.get("requires_predictability_decay_stress"))
-        or _boolish(spec.promotion_contract.get("requires_horizon_decay_curve"))
-        or _boolish(
-            spec.promotion_contract.get("requires_spread_adjusted_label_replay")
-        )
-        or "required_predictability_decay_stress" in spec.hard_vetoes
-        or "required_horizon_decay_curve" in spec.hard_vetoes
-        or "required_spread_adjusted_label_replay" in spec.hard_vetoes
-    )
-
-
-def _candidate_board_predictability_decay_summary(
-    spec: CandidateSpec, scorecard: Mapping[str, Any], *, target: Decimal
-) -> dict[str, Any]:
-    required = _candidate_spec_requires_predictability_decay_stress(spec)
-    artifact_refs = _string_list_from_value(
-        scorecard.get("predictability_decay_stress_artifact_refs")
-        or scorecard.get("predictability_decay_stress_artifact_ref")
-        or scorecard.get("alpha_decay_stress_artifact_ref")
-    )
-    available_refs = set(
-        _string_list_from_value(scorecard.get("replay_artifact_refs"))
-        or _string_list_from_value(scorecard.get("artifact_refs"))
-    )
-    horizon_count = _candidate_board_first_int_field(
-        scorecard,
-        (
-            "predictability_decay_stress_horizon_count",
-            "horizon_decay_curve_horizon_count",
-        ),
-    )
-    tight_spread_count = _candidate_board_first_int_field(
-        scorecard,
-        ("tight_spread_regime_slice_count", "tight_spread_regime_count"),
-    )
-    split_pass_rate = _decimal(
-        scorecard.get("predictability_decay_stress_split_pass_rate")
-        or scorecard.get("decay_stress_split_pass_rate")
-    )
-    best_split_share = _decimal(
-        scorecard.get("predictability_decay_stress_best_split_share")
-        or scorecard.get("decay_stress_best_split_share"),
-        default="1",
-    )
-    stress_net_pnl_per_day = _decimal(
-        scorecard.get("post_cost_net_pnl_after_predictability_decay_stress")
-        or scorecard.get("predictability_decay_stress_net_pnl_per_day")
-    )
-    min_horizon_count = _candidate_board_int_field(
-        spec.hard_vetoes, "required_min_decay_stress_horizon_count"
-    )
-    if min_horizon_count <= 0:
-        min_horizon_count = 3
-    min_tight_spread_count = _candidate_board_int_field(
-        spec.hard_vetoes, "required_min_tight_spread_regime_count"
-    )
-    if min_tight_spread_count <= 0:
-        min_tight_spread_count = 20
-    min_split_pass_rate = _decimal(
-        spec.hard_vetoes.get("required_min_decay_stress_split_pass_rate"),
-        default="0.60",
-    )
-    max_best_split_share = _decimal(
-        spec.hard_vetoes.get("required_max_decay_stress_best_split_share"),
-        default="0.35",
-    )
-    blockers: list[str] = []
-    missing_artifact_refs: list[str] = []
-    if required:
-        if not _boolish(
-            scorecard.get("predictability_decay_stress_passed")
-            or scorecard.get("alpha_decay_stress_passed")
-        ):
-            blockers.append("predictability_decay_stress_passed_failed")
-        if not artifact_refs:
-            blockers.append("predictability_decay_stress_artifact_present_failed")
-        missing_artifact_refs = [
-            ref for ref in artifact_refs if available_refs and ref not in available_refs
-        ]
-        if missing_artifact_refs:
-            blockers.append(
-                "predictability_decay_stress_artifact_ref_missing_from_bundle"
-            )
-        if not _boolish(
-            scorecard.get("horizon_decay_curve_present")
-            or scorecard.get("predictability_horizon_decay_curve_present")
-        ):
-            blockers.append("horizon_decay_curve_present_failed")
-        if not _boolish(
-            scorecard.get("spread_adjusted_label_replay_present")
-            or scorecard.get("spread_adjusted_labels_present")
-        ):
-            blockers.append("spread_adjusted_label_replay_present_failed")
-        if horizon_count < min_horizon_count:
-            blockers.append("predictability_decay_stress_horizon_count_failed")
-        if tight_spread_count < min_tight_spread_count:
-            blockers.append("tight_spread_regime_slice_count_failed")
-        if split_pass_rate < min_split_pass_rate:
-            blockers.append("predictability_decay_stress_split_pass_rate_failed")
-        if best_split_share > max_best_split_share:
-            blockers.append("predictability_decay_stress_best_split_share_failed")
-        if stress_net_pnl_per_day < target:
-            blockers.append(
-                "post_cost_net_pnl_after_predictability_decay_stress_failed"
-            )
-    return {
-        "required": required,
-        "passed": not blockers,
-        "blockers": blockers,
-        "artifact_refs": artifact_refs,
-        "missing_replay_artifact_refs": missing_artifact_refs
-        if required and artifact_refs
-        else [],
-        "horizon_count": horizon_count,
-        "min_horizon_count": min_horizon_count,
-        "tight_spread_regime_slice_count": tight_spread_count,
-        "min_tight_spread_regime_slice_count": min_tight_spread_count,
-        "split_pass_rate": str(split_pass_rate),
-        "min_split_pass_rate": str(min_split_pass_rate),
-        "best_split_share": str(best_split_share),
-        "max_best_split_share": str(max_best_split_share),
-        "stress_net_pnl_per_day": str(stress_net_pnl_per_day),
-        "target_net_pnl_per_day": str(target),
-        "source_marker": "tkan_lob_alpha_decay_arxiv_2601_02310_2026",
-    }
-
-
-def _candidate_board_scorecard_with_predictability_decay_blockers(
-    spec: CandidateSpec, scorecard: Mapping[str, Any], *, target: Decimal
-) -> tuple[dict[str, Any], dict[str, Any]]:
-    summary = _candidate_board_predictability_decay_summary(
-        spec, scorecard, target=target
-    )
-    updated_scorecard = dict(scorecard)
-    blockers = [
-        _string(blocker)
-        for blocker in cast(Sequence[Any], summary.get("blockers") or ())
-        if _string(blocker)
-    ]
-    if blockers:
-        updated_scorecard["oracle_passed"] = False
-        oracle_payload = dict(_mapping(updated_scorecard.get("profit_target_oracle")))
-        oracle_blockers = [
-            _string(blocker)
-            for blocker in cast(Sequence[Any], oracle_payload.get("blockers") or ())
-            if _string(blocker)
-        ]
-        oracle_payload["blockers"] = list(dict.fromkeys([*oracle_blockers, *blockers]))
-        updated_scorecard["profit_target_oracle"] = oracle_payload
-    return updated_scorecard, summary
-
-
-def _candidate_board_order_type_execution_quality_summary(
-    spec: CandidateSpec, scorecard: Mapping[str, Any]
-) -> dict[str, Any]:
-    required = _candidate_spec_requires_order_type_execution_quality(spec)
-    min_sample_count = _candidate_board_int_field(
-        spec.hard_vetoes, "required_min_order_type_ablation_sample_count"
-    )
-    if min_sample_count <= 0:
-        min_sample_count = 60
-    max_opportunity_cost_bps = _decimal(
-        spec.hard_vetoes.get("required_max_order_type_opportunity_cost_bps"),
-        default="8",
-    )
-    max_market_order_spread_bps = _decimal(
-        spec.hard_vetoes.get("required_max_market_order_spread_bps"),
-        default="8",
-    )
-    order_type_ablation_passed = _boolish(
-        scorecard.get("order_type_ablation_passed")
-        or scorecard.get("market_limit_execution_policy_passed")
-    )
-    raw_artifact_refs = scorecard.get(
-        "order_type_ablation_artifact_refs"
-    ) or scorecard.get("order_type_ablation_artifact_ref")
-    artifact_refs = _string_list_from_value(raw_artifact_refs)
-    if not artifact_refs and _string(raw_artifact_refs):
-        artifact_refs = [_string(raw_artifact_refs)]
-    raw_execution_artifact_refs = (
-        scorecard.get("order_type_execution_artifact_refs")
-        or scorecard.get("market_limit_order_mix_artifact_refs")
-        or scorecard.get("order_type_execution_artifact_ref")
-        or scorecard.get("market_limit_order_mix_artifact_ref")
-    )
-    execution_artifact_refs = _string_list_from_value(raw_execution_artifact_refs)
-    if not execution_artifact_refs and _string(raw_execution_artifact_refs):
-        execution_artifact_refs = [_string(raw_execution_artifact_refs)]
-    sample_count = _candidate_board_first_int_field(
-        scorecard,
-        (
-            "order_type_ablation_sample_count",
-            "market_limit_order_mix_sample_count",
-            "limit_fill_probability_sample_count",
-        ),
-    )
-    opportunity_cost_bps = _decimal(
-        scorecard.get("order_type_opportunity_cost_bps")
-        or scorecard.get("market_limit_order_mix_opportunity_cost_bps"),
-        default="999999",
-    )
-    market_order_spread_bps = _decimal(
-        scorecard.get("market_order_spread_bps")
-        or scorecard.get("market_limit_order_mix_market_spread_bps"),
-        default="999999",
-    )
-    blockers: list[str] = []
-    if required:
-        if not order_type_ablation_passed:
-            blockers.append("order_type_ablation_passed_failed")
-        if not artifact_refs:
-            blockers.append("order_type_ablation_artifact_present_failed")
-        if sample_count < min_sample_count:
-            blockers.append("order_type_ablation_sample_count_failed")
-        if not _boolish(
-            scorecard.get("market_limit_order_mix_evidence_present")
-            or scorecard.get("market_limit_order_mix_present")
-        ):
-            blockers.append("market_limit_order_mix_evidence_present_failed")
-        if not _boolish(
-            scorecard.get("limit_fill_probability_evidence_present")
-            or scorecard.get("limit_fill_probability_present")
-        ):
-            blockers.append("limit_fill_probability_evidence_present_failed")
-        if not _boolish(
-            scorecard.get("price_improvement_evidence_present")
-            or scorecard.get("route_price_improvement_evidence_present")
-        ):
-            blockers.append("price_improvement_evidence_present_failed")
-        if not _boolish(
-            scorecard.get("opportunity_cost_evidence_present")
-            or scorecard.get("order_type_opportunity_cost_evidence_present")
-        ):
-            blockers.append("opportunity_cost_evidence_present_failed")
-        if not _boolish(
-            scorecard.get("execution_shortfall_evidence_present")
-            or scorecard.get("order_type_execution_shortfall_evidence_present")
-        ):
-            blockers.append("execution_shortfall_evidence_present_failed")
-        raw_route_tca_refs = scorecard.get("route_tca_artifact_refs") or scorecard.get(
-            "route_tca_artifact_ref"
-        )
-        route_tca_refs = _string_list_from_value(raw_route_tca_refs)
-        if not route_tca_refs and _string(raw_route_tca_refs):
-            route_tca_refs = [_string(raw_route_tca_refs)]
-        if not route_tca_refs:
-            blockers.append("route_tca_evidence_present_failed")
-        if opportunity_cost_bps > max_opportunity_cost_bps:
-            blockers.append("order_type_opportunity_cost_bps_failed")
-        if market_order_spread_bps > max_market_order_spread_bps:
-            blockers.append("market_order_spread_bps_failed")
-    return {
-        "required": required,
-        "passed": not blockers,
-        "blockers": blockers,
-        "artifact_refs": artifact_refs,
-        "execution_artifact_refs": execution_artifact_refs,
-        "route_tca_artifact_refs": route_tca_refs if required else [],
-        "sample_count": sample_count,
-        "min_sample_count": min_sample_count,
-        "opportunity_cost_bps": str(opportunity_cost_bps),
-        "max_opportunity_cost_bps": str(max_opportunity_cost_bps),
-        "market_order_spread_bps": str(market_order_spread_bps),
-        "max_market_order_spread_bps": str(max_market_order_spread_bps),
-    }
-
-
-def _candidate_board_scorecard_with_order_type_blockers(
-    spec: CandidateSpec,
-    scorecard: Mapping[str, Any],
-    *,
-    replay_artifact_refs: Sequence[Any] = (),
-) -> tuple[dict[str, Any], dict[str, Any]]:
-    summary = _candidate_board_order_type_execution_quality_summary(spec, scorecard)
-    updated_scorecard = dict(scorecard)
-    blockers = [
-        _string(blocker)
-        for blocker in cast(Sequence[Any], summary.get("blockers") or ())
-        if _string(blocker)
-    ]
-    required_artifact_refs = tuple(
-        dict.fromkeys(
-            item
-            for item in (
-                *cast(Sequence[Any], summary.get("artifact_refs") or ()),
-                *cast(Sequence[Any], summary.get("execution_artifact_refs") or ()),
-                *cast(Sequence[Any], summary.get("route_tca_artifact_refs") or ()),
-            )
-            if _string(item)
-        )
-    )
-    available_artifact_refs = {
-        _string(item) for item in replay_artifact_refs if _string(item)
-    }
-    missing_artifact_refs = [
-        _string(item)
-        for item in required_artifact_refs
-        if _string(item) not in available_artifact_refs
-    ]
-    if bool(summary.get("required")) and missing_artifact_refs:
-        blockers.append("order_type_proof_artifact_ref_missing_from_bundle")
-        summary = {
-            **summary,
-            "passed": False,
-            "blockers": list(dict.fromkeys(blockers)),
-            "missing_replay_artifact_refs": missing_artifact_refs,
-            "replay_artifact_refs_checked": sorted(available_artifact_refs),
-        }
-    if blockers:
-        updated_scorecard["oracle_passed"] = False
-        oracle_payload = dict(_mapping(updated_scorecard.get("profit_target_oracle")))
-        oracle_blockers = [
-            _string(blocker)
-            for blocker in cast(Sequence[Any], oracle_payload.get("blockers") or ())
-            if _string(blocker)
-        ]
-        oracle_payload["blockers"] = list(dict.fromkeys([*oracle_blockers, *blockers]))
-        updated_scorecard["profit_target_oracle"] = oracle_payload
-    return updated_scorecard, summary
-
-
-def _candidate_spec_requires_queue_position_survival(spec: CandidateSpec) -> bool:
-    overlay_ids = set(
-        _string_list_from_value(spec.parameter_space.get("mechanism_overlay_ids"))
-    )
-    return (
-        "queue_position_survival_fill_curve" in overlay_ids
-        or _boolish(
-            spec.promotion_contract.get("requires_queue_position_survival_fill_curve")
-        )
-        or _boolish(spec.promotion_contract.get("requires_nonfill_opportunity_cost"))
-        or _boolish(spec.promotion_contract.get("requires_time_to_fill_quantiles"))
-        or "required_queue_position_survival_fill_curve" in spec.hard_vetoes
-        or "required_min_queue_position_survival_sample_count" in spec.hard_vetoes
-        or "required_max_queue_position_nonfill_opportunity_cost_bps"
-        in spec.hard_vetoes
-    )
-
-
-def _candidate_board_queue_position_survival_summary(
-    spec: CandidateSpec, scorecard: Mapping[str, Any]
-) -> dict[str, Any]:
-    required = _candidate_spec_requires_queue_position_survival(spec)
-    min_sample_count = _candidate_board_int_field(
-        spec.hard_vetoes, "required_min_queue_position_survival_sample_count"
-    )
-    if min_sample_count <= 0:
-        min_sample_count = 60
-    max_nonfill_opportunity_cost_bps = _decimal(
-        spec.hard_vetoes.get(
-            "required_max_queue_position_nonfill_opportunity_cost_bps"
-        ),
-        default="8",
-    )
-    evidence_present = _boolish(
-        scorecard.get("queue_position_survival_fill_curve_evidence_present")
-        or scorecard.get("queue_position_survival_evidence_present")
-    )
-    sample_count = _candidate_board_first_int_field(
-        scorecard,
-        (
-            "queue_position_survival_sample_count",
-            "queue_position_survival_fill_sample_count",
-        ),
-    )
-    fill_rate = _decimal(scorecard.get("queue_position_survival_fill_rate"))
-    queue_ratio_p95 = _decimal(scorecard.get("queue_position_survival_queue_ratio_p95"))
-    queue_ahead_evidence_present = _boolish(
-        scorecard.get("queue_position_survival_queue_ahead_depletion_evidence_present")
-        or scorecard.get("delay_adjusted_depth_queue_ahead_depletion_evidence_present")
-        or scorecard.get("queue_ahead_depletion_evidence_present")
-    )
-    queue_ahead_sample_count = _candidate_board_first_int_field(
-        scorecard,
-        (
-            "queue_position_survival_queue_ahead_depletion_sample_count",
-            "delay_adjusted_depth_queue_ahead_depletion_sample_count",
-            "queue_ahead_depletion_sample_count",
-        ),
-    )
-    nonfill_opportunity_cost_raw = scorecard.get(
-        "queue_position_survival_nonfill_opportunity_cost_bps"
-    ) or scorecard.get("queue_position_nonfill_opportunity_cost_bps")
-    nonfill_opportunity_cost_present = nonfill_opportunity_cost_raw not in (None, "")
-    nonfill_opportunity_cost_bps = _decimal(
-        nonfill_opportunity_cost_raw,
-        default="999999",
-    )
-    require_time_to_fill_quantiles = _boolish(
-        spec.hard_vetoes.get("required_time_to_fill_quantiles")
-    ) or _boolish(spec.promotion_contract.get("requires_time_to_fill_quantiles"))
-    time_to_fill_quantiles_present = scorecard.get("fill_time_ms_p50") not in (
-        None,
-        "",
-    ) and scorecard.get("fill_time_ms_p95") not in (None, "")
-    blockers: list[str] = []
-    if required:
-        if not evidence_present:
-            blockers.append(
-                "queue_position_survival_fill_curve_evidence_present_failed"
-            )
-        if sample_count < min_sample_count:
-            blockers.append("queue_position_survival_sample_count_failed")
-        if fill_rate <= 0:
-            blockers.append("queue_position_survival_fill_rate_failed")
-        if not queue_ahead_evidence_present:
-            blockers.append("queue_ahead_depletion_evidence_present_failed")
-        if queue_ahead_sample_count <= 0:
-            blockers.append("queue_ahead_depletion_sample_count_failed")
-        if not nonfill_opportunity_cost_present:
-            blockers.append(
-                "queue_position_survival_nonfill_opportunity_cost_present_failed"
-            )
-        elif nonfill_opportunity_cost_bps > max_nonfill_opportunity_cost_bps:
-            blockers.append(
-                "queue_position_survival_nonfill_opportunity_cost_bps_failed"
-            )
-        if require_time_to_fill_quantiles and not time_to_fill_quantiles_present:
-            blockers.append("time_to_fill_quantiles_present_failed")
-    return {
-        "required": required,
-        "passed": not blockers,
-        "blockers": blockers,
-        "evidence_present": evidence_present,
-        "sample_count": sample_count,
-        "min_sample_count": min_sample_count,
-        "fill_rate": str(fill_rate),
-        "queue_ratio_p95": str(queue_ratio_p95),
-        "queue_ahead_depletion_evidence_present": queue_ahead_evidence_present,
-        "queue_ahead_depletion_sample_count": queue_ahead_sample_count,
-        "nonfill_opportunity_cost_present": nonfill_opportunity_cost_present,
-        "nonfill_opportunity_cost_bps": str(nonfill_opportunity_cost_bps),
-        "max_nonfill_opportunity_cost_bps": str(max_nonfill_opportunity_cost_bps),
-        "time_to_fill_quantiles_required": require_time_to_fill_quantiles,
-        "time_to_fill_quantiles_present": time_to_fill_quantiles_present,
-        "source_marker": "queue_position_survival_fill_probability_arxiv_2512_05734_2025",
-    }
-
-
-def _candidate_board_scorecard_with_queue_position_survival_blockers(
-    spec: CandidateSpec, scorecard: Mapping[str, Any]
-) -> tuple[dict[str, Any], dict[str, Any]]:
-    summary = _candidate_board_queue_position_survival_summary(spec, scorecard)
-    updated_scorecard = dict(scorecard)
-    blockers = [
-        _string(blocker)
-        for blocker in cast(Sequence[Any], summary.get("blockers") or ())
-        if _string(blocker)
-    ]
-    if blockers:
-        updated_scorecard["oracle_passed"] = False
-        oracle_payload = dict(_mapping(updated_scorecard.get("profit_target_oracle")))
-        oracle_blockers = [
-            _string(blocker)
-            for blocker in cast(Sequence[Any], oracle_payload.get("blockers") or ())
-            if _string(blocker)
-        ]
-        oracle_payload["blockers"] = list(dict.fromkeys([*oracle_blockers, *blockers]))
-        updated_scorecard["profit_target_oracle"] = oracle_payload
-    return updated_scorecard, summary
-
-
-def _candidate_board_scorecard_with_rejected_signal_blockers(
-    spec: CandidateSpec, scorecard: Mapping[str, Any]
-) -> tuple[dict[str, Any], dict[str, Any]]:
-    summary = _candidate_board_rejected_signal_outcome_summary(spec, scorecard)
-    updated_scorecard = dict(scorecard)
-    blockers = [
-        _string(blocker)
-        for blocker in cast(Sequence[Any], summary.get("blockers") or ())
-        if _string(blocker)
-    ]
-    if blockers:
-        updated_scorecard["oracle_passed"] = False
-        oracle_payload = dict(_mapping(updated_scorecard.get("profit_target_oracle")))
-        oracle_blockers = [
-            _string(blocker)
-            for blocker in cast(Sequence[Any], oracle_payload.get("blockers") or ())
-            if _string(blocker)
-        ]
-        oracle_payload["blockers"] = list(dict.fromkeys([*oracle_blockers, *blockers]))
-        updated_scorecard["profit_target_oracle"] = oracle_payload
-    return updated_scorecard, summary
-
-
-def _candidate_board_evidence_lineage_summary(
-    evidence: CandidateEvidenceBundle | None,
-) -> dict[str, Any]:
-    if evidence is None:
-        return {
-            "present": False,
-            "passed": False,
-            "blockers": ["evidence_bundle_missing"],
-            "dataset_snapshot_id": "",
-            "feature_spec_hash": "",
-            "code_commit": "",
-            "replay_artifact_ref_count": 0,
-        }
-
-    blockers: list[str] = []
-    dataset_snapshot_id = _string(evidence.dataset_snapshot_id)
-    feature_spec_hash = _string(evidence.feature_spec_hash)
-    code_commit = _string(evidence.code_commit)
-    replay_artifact_refs = [
-        _string(item) for item in evidence.replay_artifact_refs if _string(item)
-    ]
-
-    if not dataset_snapshot_id:
-        blockers.append("dataset_snapshot_missing")
-    if not feature_spec_hash:
-        blockers.append("feature_spec_hash_missing")
-    if not code_commit or code_commit.lower() == "unknown":
-        blockers.append("code_commit_missing_or_unknown")
-    elif code_commit.endswith("-dirty"):
-        blockers.append("code_commit_dirty")
-    if not replay_artifact_refs:
-        blockers.append("replay_artifact_missing")
-
-    return {
-        "present": True,
-        "passed": not blockers,
-        "blockers": blockers,
-        "dataset_snapshot_id": dataset_snapshot_id,
-        "feature_spec_hash": feature_spec_hash,
-        "code_commit": code_commit,
-        "replay_artifact_ref_count": len(replay_artifact_refs),
-        "sample_replay_artifact_refs": replay_artifact_refs[:5],
-    }
-
-
-def _candidate_board_scorecard_with_lineage_blockers(
-    scorecard: Mapping[str, Any], evidence: CandidateEvidenceBundle | None
-) -> tuple[dict[str, Any], dict[str, Any]]:
-    summary = _candidate_board_evidence_lineage_summary(evidence)
-    updated_scorecard = dict(scorecard)
-    blockers = [
-        _string(blocker)
-        for blocker in cast(Sequence[Any], summary.get("blockers") or ())
-        if _string(blocker)
-    ]
-    if blockers:
-        updated_scorecard["oracle_passed"] = False
-        oracle_payload = dict(_mapping(updated_scorecard.get("profit_target_oracle")))
-        oracle_blockers = [
-            _string(blocker)
-            for blocker in cast(Sequence[Any], oracle_payload.get("blockers") or ())
-            if _string(blocker)
-        ]
-        oracle_payload["blockers"] = list(dict.fromkeys([*oracle_blockers, *blockers]))
-        updated_scorecard["profit_target_oracle"] = oracle_payload
-    return updated_scorecard, summary
-
-
-def _candidate_board_replay_window_coverage_summary(
-    scorecard: Mapping[str, Any],
-) -> dict[str, Any]:
-    required = _boolish(scorecard.get("target_met")) or _boolish(
-        scorecard.get("oracle_passed")
-    )
-    replay_lineage = _mapping(scorecard.get("replay_lineage"))
-    coverage = _mapping(scorecard.get("replay_window_coverage"))
-    expected_windows = _string_list_from_value(
-        coverage.get("expected_windows") or replay_lineage.get("expected_windows")
-    )
-    present_windows = _string_list_from_value(
-        coverage.get("present_windows") or replay_lineage.get("present_windows")
-    )
-    missing_windows = _string_list_from_value(
-        coverage.get("missing_windows") or replay_lineage.get("missing_windows")
-    )
-    if (
-        _candidate_board_int_field(scorecard, "double_oos_independent_window_count")
-        >= 2
-        and _SECOND_OOS_WINDOW_ID not in expected_windows
-    ):
-        expected_windows.append(_SECOND_OOS_WINDOW_ID)
-    for required_window in ("train", "holdout", "full_window"):
-        if required_window not in expected_windows:
-            expected_windows.append(required_window)
-
-    blockers: list[str] = []
-    lineage_hash = _string(replay_lineage.get("lineage_hash"))
-    coverage_hash = _string(coverage.get("lineage_hash"))
-    if required:
-        if not replay_lineage:
-            blockers.append("replay_lineage_missing")
-        if not coverage:
-            blockers.append("replay_window_coverage_missing")
-        if not lineage_hash:
-            blockers.append("replay_lineage_hash_missing")
-        if not coverage_hash:
-            blockers.append("replay_window_coverage_hash_missing")
-        if lineage_hash and coverage_hash and lineage_hash != coverage_hash:
-            blockers.append("replay_lineage_hash_mismatch")
-        missing_required_windows = [
-            window_id
-            for window_id in expected_windows
-            if window_id not in present_windows or window_id in missing_windows
-        ]
-        if missing_required_windows:
-            blockers.append("replay_window_missing")
-    else:
-        missing_required_windows = []
-
-    return {
-        "required": required,
-        "passed": not blockers,
-        "blockers": blockers,
-        "lineage_hash": lineage_hash,
-        "coverage_hash": coverage_hash,
-        "expected_windows": expected_windows,
-        "present_windows": present_windows,
-        "missing_windows": missing_windows,
-        "missing_required_windows": missing_required_windows,
-    }
-
-
-def _candidate_board_market_impact_proof_summary(
-    scorecard: Mapping[str, Any],
-) -> dict[str, Any]:
-    components = _mapping(scorecard.get("market_impact_stress_components"))
-    source_marker = _string(components.get("source_marker"))
-    source_marker_candidates = [
-        *_string_list_from_value(scorecard.get("market_impact_stress_source_markers")),
-        *_string_list_from_value(components.get("source_markers")),
-        *([source_marker] if source_marker else []),
-    ]
-    source_markers = sorted({marker for marker in source_marker_candidates if marker})
-    model = _string(
-        scorecard.get("nonlinear_market_impact_stress_model")
-        or scorecard.get("market_impact_stress_model")
-    )
-    artifact_ref = _string(
-        scorecard.get("market_impact_stress_artifact_ref")
-        or scorecard.get("impact_stress_artifact_ref")
-        or scorecard.get("cost_shock_artifact_ref")
-    )
-    cost_bps = _candidate_board_decimal_field(
-        scorecard, "nonlinear_market_impact_stress_cost_bps"
-    ) or _candidate_board_decimal_field(scorecard, "market_impact_stress_cost_bps")
-    net_pnl_per_day = _candidate_board_decimal_field(
-        scorecard, "nonlinear_market_impact_stress_net_pnl_per_day"
-    ) or _candidate_board_decimal_field(
-        scorecard, "market_impact_stress_net_pnl_per_day"
-    )
-    nonlinear_passed = _boolish(
-        scorecard.get("nonlinear_market_impact_stress_passed")
-        or scorecard.get("market_impact_stress_passed")
-    )
-    blockers: list[str] = []
-    if not artifact_ref:
-        blockers.append("market_impact_stress_artifact_missing")
-    if not model:
-        blockers.append("market_impact_stress_model_missing")
-    if not cost_bps or _decimal(cost_bps) <= 0:
-        blockers.append("market_impact_stress_cost_bps_missing")
-    if not net_pnl_per_day:
-        blockers.append("market_impact_stress_net_pnl_missing")
-    if not source_marker:
-        blockers.append("nonlinear_market_impact_components_missing")
-    if model and artifact_ref and cost_bps and not nonlinear_passed:
-        blockers.append("nonlinear_market_impact_stress_failed")
-    if not scorecard:
-        state = "not_replayed"
-    elif blockers:
-        state = "blocked"
-    else:
-        state = "passed"
-    return {
-        "state": state,
-        "passed": state == "passed",
-        "model": model,
-        "cost_bps": cost_bps,
-        "net_pnl_per_day": net_pnl_per_day,
-        "artifact_ref": artifact_ref,
-        "component_source_marker": source_marker,
-        "source_markers": source_markers,
-        "selected_component_model": _string(components.get("selected_model")),
-        "selected_component_cost_bps": _string(components.get("selected_cost_bps")),
-        "blockers": blockers,
-        "source_marker": "realistic_market_impact_arxiv_2603_29086_2026",
-    }
-
-
-def _candidate_board_regime_specialist_summary(
-    spec: CandidateSpec,
-    scorecard: Mapping[str, Any],
-) -> dict[str, Any]:
-    required_rate = _decimal(
-        spec.hard_vetoes.get("required_min_regime_slice_pass_rate")
-        or spec.promotion_contract.get("required_min_regime_slice_pass_rate")
-    )
-    observed_rate = _decimal(scorecard.get("regime_slice_pass_rate"))
-    source_claims = _list_of_mappings(spec.feature_contract.get("source_claims"))
-    regime_claim_ids = [
-        _string(claim.get("claim_id"))
-        for claim in source_claims
-        if _string(claim.get("claim_type"))
-        in {"market_regime", "validation_requirement", "risk_constraint"}
-    ]
-    blockers: list[str] = []
-    if scorecard and required_rate > 0 and "regime_slice_pass_rate" not in scorecard:
-        blockers.append("regime_slice_pass_rate_missing")
-    if scorecard and required_rate > 0 and observed_rate < required_rate:
-        blockers.append("regime_slice_pass_rate_below_specialist_threshold")
-    if not scorecard:
-        state = "not_replayed"
-    elif blockers:
-        state = "blocked"
-    else:
-        state = "passed"
-    return {
-        "state": state,
-        "passed": state == "passed",
-        "regime_slice_pass_rate": str(observed_rate) if scorecard else "",
-        "required_min_regime_slice_pass_rate": str(required_rate)
-        if required_rate > 0
-        else "",
-        "regime_claim_ids": [claim_id for claim_id in regime_claim_ids if claim_id],
-        "blockers": blockers,
-        "source_markers": [
-            "risk_sensitive_specialist_routing_arxiv_2604_10402_2026",
-            "validated_vvg_classifier_arxiv_2605_11423_2026",
-        ],
-    }
-
-
-def _candidate_board_scorecard_with_replay_window_blockers(
-    scorecard: Mapping[str, Any],
-) -> tuple[dict[str, Any], dict[str, Any]]:
-    summary = _candidate_board_replay_window_coverage_summary(scorecard)
-    updated_scorecard = dict(scorecard)
-    blockers = [
-        _string(blocker)
-        for blocker in cast(Sequence[Any], summary.get("blockers") or ())
-        if _string(blocker)
-    ]
-    if blockers:
-        updated_scorecard["oracle_passed"] = False
-        oracle_payload = dict(_mapping(updated_scorecard.get("profit_target_oracle")))
-        oracle_blockers = [
-            _string(blocker)
-            for blocker in cast(Sequence[Any], oracle_payload.get("blockers") or ())
-            if _string(blocker)
-        ]
-        oracle_payload["blockers"] = list(dict.fromkeys([*oracle_blockers, *blockers]))
-        updated_scorecard["profit_target_oracle"] = oracle_payload
-    return updated_scorecard, summary
-
-
-def _candidate_board_scorecard_with_evidence_blockers(
-    scorecard: Mapping[str, Any], evidence: CandidateEvidenceBundle | None
-) -> dict[str, Any]:
-    if evidence is None:
-        return dict(scorecard)
-    blockers = [
-        _string(blocker)
-        for blocker in (
-            *evidence_bundle_blockers(evidence),
-            *cast(
-                Sequence[Any],
-                evidence.promotion_readiness.get("blockers") or (),
-            ),
-        )
-        if _string(blocker)
-    ]
-    if not blockers:
-        return dict(scorecard)
-    updated_scorecard = dict(scorecard)
-    updated_scorecard["oracle_passed"] = False
-    oracle_payload = dict(_mapping(updated_scorecard.get("profit_target_oracle")))
-    oracle_blockers = [
-        _string(blocker)
-        for blocker in cast(Sequence[Any], oracle_payload.get("blockers") or ())
-        if _string(blocker)
-    ]
-    oracle_payload["blockers"] = list(dict.fromkeys([*oracle_blockers, *blockers]))
-    updated_scorecard["profit_target_oracle"] = oracle_payload
-    return updated_scorecard
-
-
-def _candidate_board_blockers(
-    *,
-    selected_for_replay: bool,
-    evidence: CandidateEvidenceBundle | None,
-    scorecard: Mapping[str, Any],
-) -> list[str]:
-    blockers = list(_oracle_blockers(scorecard))
-    if not selected_for_replay:
-        blockers.append("not_selected_for_replay")
-    elif evidence is None:
-        blockers.append("replay_evidence_missing")
-    if evidence is not None and not _boolish(scorecard.get("target_met")):
-        blockers.append("target_net_pnl_per_day_not_met")
-    if (
-        evidence is not None
-        and _boolish(scorecard.get("target_met"))
-        and not _boolish(scorecard.get("oracle_passed"))
-        and not blockers
-    ):
-        blockers.append("profit_target_oracle_failed")
-    return list(dict.fromkeys(blockers))
-
-
-def _candidate_board_status(
-    *,
-    selected_for_replay: bool,
-    evidence: CandidateEvidenceBundle | None,
-    scorecard: Mapping[str, Any],
-    in_best_portfolio: bool,
-    portfolio_oracle_passed: bool,
-) -> str:
-    if in_best_portfolio and portfolio_oracle_passed:
-        return "portfolio_component_passed_oracle"
-    if evidence is not None and _boolish(scorecard.get("oracle_passed")):
-        return "candidate_oracle_passed"
-    if evidence is not None and _boolish(scorecard.get("target_met")):
-        return "blocked_by_oracle"
-    if evidence is not None:
-        return "replayed_below_target"
-    if selected_for_replay:
-        return "selected_pending_replay_evidence"
-    return "research_ranked_not_replayed"
-
-
-def _candidate_board_activity_count(row: Mapping[str, Any]) -> int:
-    return max(
-        _candidate_board_int_field(row, "decision_count"),
-        _candidate_board_int_field(row, "submitted_order_count"),
-        _candidate_board_int_field(row, "filled_order_count"),
-        _candidate_board_int_field(row, "executable_replay_order_count"),
-    )
-
-
-def _candidate_board_oracle_blocker_count(row: Mapping[str, Any]) -> int:
-    blockers = row.get("blockers")
-    if not isinstance(blockers, Sequence) or isinstance(blockers, str):
-        return 0
-    return sum(
-        1
-        for blocker in cast(Sequence[Any], blockers)
-        if _string(blocker).endswith("_failed")
-    )
-
-
-def _candidate_board_net_pnl(row: Mapping[str, Any]) -> Decimal:
-    return _decimal(row.get("net_pnl_per_day"))
-
-
-def _candidate_board_lower_bound_net_pnl(row: Mapping[str, Any]) -> Decimal:
-    lower_bound = deployable_lower_bound_net_pnl_per_day(row)
-    if lower_bound is None:
-        return Decimal("0")
-    if (
-        deployable_lower_bound_missing_count(row) > 0
-        or deployable_proof_failed_gate_count(row) > 0
-    ):
-        return Decimal("0")
-    return lower_bound
-
-
-def _candidate_board_target_progress_ratio(
-    *,
-    lower_bound: Decimal,
-    target: Decimal,
-) -> Decimal:
-    if target <= 0:
-        return Decimal("0")
-    return (lower_bound / target).quantize(Decimal("0.0001"))
-
-
-def _candidate_board_required_notional_repair_scale(
-    *,
-    lower_bound: Decimal,
-    target: Decimal,
-) -> Decimal:
-    if target <= 0 or lower_bound <= 0:
-        return Decimal("0")
-    return max(Decimal("1"), target / lower_bound).quantize(Decimal("0.0001"))
-
-
-def _candidate_board_best_executed_candidate(
-    rows: Sequence[Mapping[str, Any]],
-) -> dict[str, Any] | None:
-    executed_rows = [
-        dict(row) for row in rows if _candidate_board_activity_count(row) > 0
-    ]
-    if not executed_rows:
-        return None
-    executed_rows.sort(
-        key=lambda row: (
-            bool(row.get("oracle_passed")),
-            bool(row.get("target_met")),
-            _candidate_board_activity_count(row),
-            _candidate_board_net_pnl(row),
-            -_rank_sort_value(row.get("rank")),
-            _string(row.get("candidate_spec_id")),
-        ),
-        reverse=True,
-    )
-    return executed_rows[0]
-
-
-def _candidate_board_closest_promotion_candidate(
-    rows: Sequence[Mapping[str, Any]],
-) -> dict[str, Any] | None:
-    replayed_rows = [dict(row) for row in rows if bool(row.get("has_replay_evidence"))]
-    if not replayed_rows:
-        return None
-    replayed_rows.sort(
-        key=lambda row: (
-            bool(row.get("oracle_passed")),
-            bool(row.get("target_met")),
-            -_candidate_board_oracle_blocker_count(row),
-            _candidate_board_activity_count(row),
-            _candidate_board_net_pnl(row),
-            -_rank_sort_value(row.get("rank")),
-            _string(row.get("candidate_spec_id")),
-        ),
-        reverse=True,
-    )
-    return replayed_rows[0]
-
-
-def _candidate_board_paper_probation_admission_blockers(
-    row: Mapping[str, Any],
-) -> list[str]:
-    blockers: list[str] = []
-    exact_replay_ledger_refs = _candidate_board_exact_replay_ledger_refs(row)
-    if not exact_replay_ledger_refs:
-        blockers.append("paper_probation_exact_replay_ledger_artifact_missing")
-    if _candidate_board_int_field(row, "exact_replay_ledger_artifact_row_count") <= 0:
-        blockers.append("paper_probation_exact_replay_ledger_row_count_missing")
-    if _candidate_board_int_field(row, "exact_replay_ledger_artifact_fill_count") <= 0:
-        blockers.append("paper_probation_exact_replay_ledger_fill_count_missing")
-    window_start, window_end = _candidate_board_runtime_window_bounds(row)
-    if not window_start or not window_end:
-        blockers.append("paper_probation_runtime_window_bounds_missing")
-    return blockers
-
-
-def _paper_probation_candidate_payload(
-    *,
-    candidate: Mapping[str, Any],
-    target: Decimal,
-    rank: int,
-) -> dict[str, Any]:
-    payload = dict(candidate)
-    lower_bound = _candidate_board_lower_bound_net_pnl(payload)
-    target_shortfall = max(target - lower_bound, Decimal("0"))
-    target_progress_ratio = _candidate_board_target_progress_ratio(
-        lower_bound=lower_bound,
-        target=target,
-    )
-    required_notional_repair_scale = _candidate_board_required_notional_repair_scale(
-        lower_bound=lower_bound,
-        target=target,
-    )
-    deployable_missing_count = deployable_lower_bound_missing_count(payload)
-    deployable_failed_gate_count = deployable_proof_failed_gate_count(payload)
-    final_promotion_blockers = [
-        _string(blocker)
-        for blocker in cast(Sequence[Any], payload.get("blockers") or ())
-        if _string(blocker)
-    ]
-    if not final_promotion_blockers:
-        final_promotion_blockers.append("final_promotion_requires_runtime_governance")
-    live_paper_evidence_requirements = list(
-        _PAPER_PROBATION_LIVE_PAPER_EVIDENCE_REQUIREMENTS
-    )
-    safe_evidence_collection_path = list(_PAPER_PROBATION_SAFE_EVIDENCE_COLLECTION_PATH)
-    if bool(payload.get("required_seed_model_family_robustness")):
-        live_paper_evidence_requirements.extend(
-            (
-                "seed_robustness_replay_grid",
-                "model_family_robustness_replay_grid",
-            )
-        )
-        safe_evidence_collection_path.extend(
-            (
-                "run_offline_seed_model_family_robustness_grid_before_paper_probation",
-                "attach_seed_model_family_robustness_artifact_to_evidence_bundle",
-            )
-        )
-    payload.update(
-        {
-            "candidate_selection": "oracle_recommended_paper_probation",
-            "paper_probation_tier": "paper_probation",
-            "paper_probation_rank": rank,
-            "paper_probation_authorized": True,
-            "probation_allowed": True,
-            "paper_probation_authorization_scope": "evidence_collection_only",
-            "evidence_collection_stage": "paper",
-            "probation_reason": "runtime_evidence_collection_only",
-            "selection_reason": "target_met_but_oracle_blocked"
-            if bool(payload.get("target_met"))
-            else "closest_lower_bound_economics_below_target",
-            "selected_by": "candidate_board_paper_probation_candidate",
-            "probation_lower_bound_net_pnl_per_day": _decimal_payload(lower_bound),
-            "probation_target_shortfall": _decimal_payload(target_shortfall),
-            "probation_target_progress_ratio": _decimal_payload(target_progress_ratio),
-            "required_notional_repair_scale_to_target": _decimal_payload(
-                required_notional_repair_scale
-            ),
-            "required_notional_repair_scale_authority": (
-                "linear_notional_sizing_estimate_for_repair_only_not_capital_authority"
-            ),
-            "live_paper_evidence_requirements": list(
-                dict.fromkeys(live_paper_evidence_requirements)
-            ),
-            "safe_evidence_collection_path": list(
-                dict.fromkeys(safe_evidence_collection_path)
-            ),
-            "live_capital_authorized": False,
-            "final_promotion_requires_live_paper_runtime_proof": True,
-            "deployable_lower_bound_missing_count": deployable_missing_count,
-            "deployable_lower_bound_failed_gate_count": deployable_failed_gate_count,
-            "promotion_allowed": False,
-            "promotion_gate": "existing_runtime_governance_fail_closed",
-            "final_promotion_authorized": False,
-            "final_promotion_allowed": False,
-            "promotion_blockers": list(dict.fromkeys(final_promotion_blockers)),
-            "final_promotion_blockers": list(dict.fromkeys(final_promotion_blockers)),
-        }
-    )
-    return payload
-
-
-def _candidate_board_paper_probation_candidates(
-    rows: Sequence[Mapping[str, Any]],
-    *,
-    target: Decimal,
-    limit: int = 1,
-) -> tuple[dict[str, Any], ...]:
-    candidates = [
-        dict(row)
-        for row in rows
-        if bool(row.get("has_replay_evidence"))
-        and not bool(row.get("oracle_passed"))
-        and _candidate_board_activity_count(row) > 0
-        and _string(row.get("candidate_id"))
-        and _string(row.get("hypothesis_id"))
-        and _string(row.get("runtime_family"))
-        and _string(row.get("runtime_strategy_name"))
-    ]
-    if not candidates:
-        return ()
-
-    admitted_candidates = [
-        candidate
-        for candidate in candidates
-        if not _candidate_board_paper_probation_admission_blockers(candidate)
-        and _candidate_board_lower_bound_net_pnl(candidate) > 0
-    ]
-    if not admitted_candidates:
-        return ()
-
-    def rank(row: Mapping[str, Any]) -> tuple[Any, ...]:
-        lower_bound = _candidate_board_lower_bound_net_pnl(row)
-        target_shortfall = max(target - lower_bound, Decimal("0"))
-        return (
-            lower_bound > 0,
-            _candidate_board_int_field(row, "filled_order_count") > 0,
-            _candidate_board_int_field(row, "submitted_order_count") > 0,
-            -target_shortfall,
-            lower_bound,
-            _decimal(row.get("active_day_ratio")),
-            _decimal(row.get("positive_day_ratio")),
-            -_decimal(row.get("best_day_share"), default="1"),
-            -_decimal(row.get("worst_day_loss")),
-            -_candidate_board_oracle_blocker_count(row),
-            _candidate_board_activity_count(row),
-            -_rank_sort_value(row.get("rank")),
-            _string(row.get("candidate_spec_id")),
-        )
-
-    admitted_candidates.sort(key=rank, reverse=True)
-    selected = admitted_candidates[: max(1, int(limit))]
-    return tuple(
-        _paper_probation_candidate_payload(
-            candidate=candidate,
-            target=target,
-            rank=index,
-        )
-        for index, candidate in enumerate(selected, start=1)
-    )
-
-
-def _candidate_board_paper_probation_candidate(
-    rows: Sequence[Mapping[str, Any]],
-    *,
-    target: Decimal,
-) -> dict[str, Any] | None:
-    candidates = _candidate_board_paper_probation_candidates(
-        rows,
-        target=target,
-        limit=1,
-    )
-    return candidates[0] if candidates else None
-
-
-def _candidate_board_status_digest(rows: Sequence[Mapping[str, Any]]) -> str:
-    digest_rows = [
-        {
-            "candidate_spec_id": _string(row.get("candidate_spec_id")),
-            "candidate_id": _string(row.get("candidate_id")),
-            "rank": _rank_sort_value(row.get("rank")),
-            "status": _string(row.get("status")),
-            "target_met": bool(row.get("target_met")),
-            "oracle_passed": bool(row.get("oracle_passed")),
-            "activity_count": _candidate_board_activity_count(row),
-            "market_impact_proof_state": _string(
-                _mapping(row.get("market_impact_proof")).get("state")
-            ),
-            "regime_specialist_state": _string(
-                _mapping(row.get("regime_specialist_validation")).get("state")
-            ),
-            "blockers": [
-                _string(blocker)
-                for blocker in cast(Sequence[Any], row.get("blockers") or ())
-            ],
-        }
-        for row in rows
-    ]
-    encoded = json.dumps(digest_rows, sort_keys=True, separators=(",", ":"))
-    return hashlib.sha256(encoded.encode("utf-8")).hexdigest()
-
-
-def _candidate_board_double_oos_summary(
-    rows: Sequence[Mapping[str, Any]],
-) -> dict[str, Any]:
-    replayed_rows = [row for row in rows if bool(row.get("has_replay_evidence"))]
-    artifact_rows = [
-        row
-        for row in replayed_rows
-        if bool(_string(row.get("double_oos_artifact_ref")))
-    ]
-    passed_rows = [row for row in replayed_rows if bool(row.get("double_oos_passed"))]
-    missing_artifact_rows = [
-        row
-        for row in replayed_rows
-        if not bool(_string(row.get("double_oos_artifact_ref")))
-    ]
-    independent_window_counts = [
-        _candidate_board_int_field(row, "double_oos_independent_window_count")
-        for row in replayed_rows
-    ]
-    blockers: list[str] = []
-    if replayed_rows and missing_artifact_rows:
-        blockers.append("double_oos_artifact_missing_for_replayed_candidates")
-    if replayed_rows and not passed_rows:
-        blockers.append("no_candidate_passed_double_oos")
-    return {
-        "replayed_candidate_count": len(replayed_rows),
-        "artifact_candidate_count": len(artifact_rows),
-        "passed_candidate_count": len(passed_rows),
-        "missing_artifact_candidate_count": len(missing_artifact_rows),
-        "max_independent_window_count": max(independent_window_counts, default=0),
-        "blockers": blockers,
-    }
-
-
-def _candidate_board_portfolio_promotion_subject(
-    *,
-    portfolio: PortfolioCandidateSpec | None,
-    portfolio_payload: Mapping[str, Any],
-    portfolio_scorecard: Mapping[str, Any],
-    promotion_readiness: Mapping[str, Any],
-    runtime_closure: Mapping[str, Any],
-    rows: Sequence[Mapping[str, Any]],
-) -> dict[str, Any] | None:
-    if portfolio is None:
-        return None
-    source_candidate_ids = [
-        _string(candidate_id)
-        for candidate_id in cast(Sequence[Any], portfolio.source_candidate_ids)
-        if _string(candidate_id)
-    ]
-    sleeve_candidate_spec_ids = [
-        _string(sleeve.get("candidate_spec_id"))
-        for sleeve in _list_of_mappings(portfolio_payload.get("sleeves"))
-        if _string(sleeve.get("candidate_spec_id"))
-    ]
-    readiness_blockers = [
-        _string(blocker)
-        for blocker in cast(Sequence[Any], promotion_readiness.get("blockers") or ())
-        if _string(blocker)
-    ]
-    blockers = list(
-        dict.fromkeys([*_oracle_blockers(portfolio_scorecard), *readiness_blockers])
-    )
-    return {
-        "type": "portfolio",
-        "portfolio_candidate_id": _string(
-            portfolio_payload.get("portfolio_candidate_id")
-        ),
-        "source_candidate_ids": source_candidate_ids,
-        "sleeve_candidate_spec_ids": sleeve_candidate_spec_ids,
-        "oracle_passed": _boolish(portfolio_scorecard.get("oracle_passed")),
-        "target_met": _boolish(portfolio_scorecard.get("target_met")),
-        "net_pnl_per_day": _candidate_board_decimal_field(
-            portfolio_scorecard, "net_pnl_per_day"
-        )
-        or _candidate_board_decimal_field(
-            portfolio_scorecard, "portfolio_post_cost_net_pnl_per_day"
-        ),
-        "market_impact_proof": _candidate_board_market_impact_proof_summary(
-            portfolio_scorecard
-        ),
-        "runtime_closure_status": _string(runtime_closure.get("status")),
-        "promotion_readiness_status": _string(promotion_readiness.get("status")),
-        "promotable": _boolish(promotion_readiness.get("promotable")),
-        "component_rows": [
-            {
-                "candidate_spec_id": _string(row.get("candidate_spec_id")),
-                "candidate_id": _string(row.get("candidate_id")),
-                "status": _string(row.get("status")),
-                "oracle_passed": _boolish(row.get("oracle_passed")),
-                "target_met": _boolish(row.get("target_met")),
-                "market_impact_proof_state": _string(
-                    _mapping(row.get("market_impact_proof")).get("state")
-                ),
-                "regime_specialist_state": _string(
-                    _mapping(row.get("regime_specialist_validation")).get("state")
-                ),
-            }
-            for row in rows
-            if _string(row.get("candidate_spec_id")) in set(sleeve_candidate_spec_ids)
-        ],
-        "blockers": blockers,
-    }
-
-
-def _candidate_board_hypothesis_manifest_ref(hypothesis_id: str | None) -> str:
-    normalized = _string(hypothesis_id).lower().replace("_", "-")
-    if not normalized:
-        return ""
-    return f"config/trading/hypotheses/{normalized}.json"
-
-
-def _candidate_board_runtime_window_bounds(
-    scorecard: Mapping[str, Any],
-) -> tuple[str, str]:
-    for source in (
-        scorecard,
-        _mapping(scorecard.get("replay_window_spec")),
-        _mapping(scorecard.get("candidate_evaluation_key_payload")).get(
-            "replay_window_spec"
-        ),
-    ):
-        mapping = _mapping(source)
-        start = _string(
-            mapping.get("runtime_window_start")
-            or mapping.get("window_start")
-            or mapping.get("full_window_start")
-            or mapping.get("full_window_start_date")
-        )
-        end = _string(
-            mapping.get("runtime_window_end")
-            or mapping.get("window_end")
-            or mapping.get("full_window_end")
-            or mapping.get("full_window_end_date")
-        )
-        if start and end:
-            return start, end
-
-    replay_lineage = _mapping(scorecard.get("replay_lineage"))
-    windows = _mapping(replay_lineage.get("windows"))
-    for window_id in ("full_window", "holdout", "train"):
-        window = _mapping(windows.get(window_id))
-        start = _string(window.get("start_date") or window.get("window_start"))
-        end = _string(window.get("end_date") or window.get("window_end"))
-        if start and end:
-            return start, end
-    return "", ""
-
-
-def _candidate_board_date_only(value: str) -> date | None:
-    text = _string(value)
-    if not text or "T" in text or " " in text:
-        return None
-    try:
-        return date.fromisoformat(text)
-    except ValueError:
-        return None
-
-
-def _candidate_board_regular_session_bound(value: str, *, end: bool) -> str:
-    text = _string(value)
-    trading_day = _candidate_board_date_only(text)
-    if trading_day is None:
-        return text
-    session_time = (
-        _CANDIDATE_BOARD_RUNTIME_SESSION_CLOSE
-        if end
-        else _CANDIDATE_BOARD_RUNTIME_SESSION_OPEN
-    )
-    return (
-        datetime.combine(
-            trading_day,
-            session_time,
-            tzinfo=_CANDIDATE_BOARD_RUNTIME_SESSION_TZ,
-        )
-        .astimezone(UTC)
-        .isoformat()
-    )
-
-
-def _candidate_board_runtime_window_import_bounds(
-    row: Mapping[str, Any],
-) -> tuple[str, str]:
-    window_start = _string(row.get("runtime_window_start") or row.get("window_start"))
-    window_end = _string(row.get("runtime_window_end") or row.get("window_end"))
-    if not window_start or not window_end:
-        window_start, window_end = _candidate_board_runtime_window_bounds(row)
-    return (
-        _candidate_board_regular_session_bound(window_start, end=False),
-        _candidate_board_regular_session_bound(window_end, end=True),
-    )
-
-
-def _candidate_board_exact_replay_ledger_refs(
-    row: Mapping[str, Any],
-) -> tuple[str, ...]:
-    refs: list[str] = []
-    for key in ("exact_replay_ledger_artifact_ref",):
-        ref = _string(row.get(key))
-        if ref:
-            refs.append(ref)
-    for key in ("exact_replay_ledger_artifact_refs",):
-        raw_refs = row.get(key)
-        if isinstance(raw_refs, str):
-            ref = _string(raw_refs)
-            if ref:
-                refs.append(ref)
-            continue
-        for raw_ref in cast(Sequence[Any], raw_refs or ()):
-            ref = _string(raw_ref)
-            if ref:
-                refs.append(ref)
-    for raw_ref in cast(Sequence[Any], row.get("replay_artifact_refs") or ()):
-        ref = _string(raw_ref)
-        ref_lower = ref.lower()
-        if ref and (
-            "exact-replay-ledger" in ref_lower or "exact_replay_ledger" in ref_lower
-        ):
-            refs.append(ref)
-    return tuple(dict.fromkeys(refs))
-
-
-def _candidate_board_runtime_import_args(target: Mapping[str, Any]) -> list[str]:
-    args = [
-        "uv",
-        "run",
-        "--frozen",
-        "python",
-        "scripts/import_hypothesis_runtime_windows.py",
-        "--run-id",
-        _string(target.get("run_id")) or "candidate-board-runtime-window-import",
-        "--candidate-id",
-        _string(target.get("candidate_id")),
-        "--hypothesis-id",
-        _string(target.get("hypothesis_id")),
-        "--observed-stage",
-        _string(target.get("observed_stage")) or "paper",
-        "--strategy-family",
-        _string(target.get("strategy_family")),
-        "--strategy-name",
-        _string(target.get("strategy_name")),
-        "--account-label",
-        _string(target.get("account_label")),
-        "--window-start",
-        _string(target.get("window_start")),
-        "--window-end",
-        _string(target.get("window_end")),
-        "--source-kind",
-        _string(target.get("source_kind")),
-        "--source-manifest-ref",
-        _string(target.get("source_manifest_ref")),
-        "--dataset-snapshot-ref",
-        _string(target.get("dataset_snapshot_ref")),
-    ]
-    artifact_refs = (
-        target.get("exact_replay_ledger_artifact_refs")
-        or target.get("runtime_ledger_artifact_refs")
-        or ()
-    )
-    for artifact_ref in cast(Sequence[Any], artifact_refs):
-        ref = _string(artifact_ref)
-        if ref:
-            args.extend(["--artifact-ref", ref])
-    target_metadata = _mapping(target.get("target_metadata"))
-    if target_metadata:
-        args.extend(
-            [
-                "--target-metadata-json",
-                json.dumps(
-                    target_metadata,
-                    sort_keys=True,
-                    separators=(",", ":"),
-                    default=str,
-                ),
-            ]
-        )
-    args.append("--json")
-    return args
-
-
-def _candidate_board_runtime_window_import_plan(
-    *,
-    rows: Sequence[Mapping[str, Any]],
-    paper_probation_candidate: Mapping[str, Any] | None,
-    paper_probation_candidates: Sequence[Mapping[str, Any]] = (),
-    promotion_subject: Mapping[str, Any] | None = None,
-) -> dict[str, Any]:
-    source_rows: list[Mapping[str, Any]] = []
-    if paper_probation_candidate is not None:
-        source_rows.append(paper_probation_candidate)
-    source_rows.extend(paper_probation_candidates)
-
-    if promotion_subject is not None and _boolish(promotion_subject.get("target_met")):
-        portfolio_spec_ids = {
-            _string(candidate_spec_id)
-            for candidate_spec_id in cast(
-                Sequence[Any],
-                promotion_subject.get("sleeve_candidate_spec_ids") or (),
-            )
-            if _string(candidate_spec_id)
-        }
-        for row in rows:
-            if _string(row.get("candidate_spec_id")) in portfolio_spec_ids:
-                source_rows.append(row)
-
-    targets: list[dict[str, Any]] = []
-    blockers: list[dict[str, Any]] = []
-    seen_keys: set[tuple[str, str, str]] = set()
-    for row in source_rows:
-        candidate_id = _string(row.get("candidate_id"))
-        hypothesis_id = _string(row.get("hypothesis_id"))
-        strategy_family = _string(row.get("runtime_family"))
-        strategy_name = _string(row.get("runtime_strategy_name"))
-        candidate_spec_id = _string(row.get("candidate_spec_id"))
-        exact_replay_ledger_artifact_refs = _candidate_board_exact_replay_ledger_refs(
-            row
-        )
-        exact_replay_ledger_artifact_ref = _string(
-            row.get("exact_replay_ledger_artifact_ref")
-        )
-        if not exact_replay_ledger_artifact_ref:
-            exact_replay_ledger_artifact_ref = next(
-                (
-                    ref
-                    for ref in exact_replay_ledger_artifact_refs
-                    if "exact" in ref.lower()
-                ),
-                "",
-            )
-        window_start, window_end = _candidate_board_runtime_window_import_bounds(row)
-        account_label = _string(row.get("account_label"))
-        if not account_label and exact_replay_ledger_artifact_refs:
-            account_label = "TORGHUT_REPLAY"
-        missing_fields = [
-            field
-            for field, value in (
-                ("candidate_id", candidate_id),
-                ("hypothesis_id", hypothesis_id),
-                ("runtime_family", strategy_family),
-                ("runtime_strategy_name", strategy_name),
-                ("window_start", window_start),
-                ("window_end", window_end),
-                ("account_label", account_label),
-            )
-            if not value
-        ]
-        if missing_fields:
-            blockers.append(
-                {
-                    "blocker": "runtime_window_import_target_incomplete",
-                    "candidate_spec_id": candidate_spec_id,
-                    "candidate_id": candidate_id,
-                    "missing_fields": missing_fields,
-                }
-            )
-            continue
-        dedupe_key = (hypothesis_id, candidate_id, strategy_name)
-        if dedupe_key in seen_keys:
-            continue
-        seen_keys.add(dedupe_key)
-        runtime_ledger_lineage_handoff = _mapping(
-            row.get("runtime_ledger_lineage_materialization_handoff")
-        )
-        runtime_ledger_required_materialized_artifacts = (
-            _candidate_board_runtime_ledger_required_materialized_artifacts(
-                runtime_ledger_lineage_handoff
-            )
-        )
-        target = {
-            "run_id": _string(row.get("epoch_id"))
-            or "candidate-board-runtime-window-import",
-            "candidate_spec_id": candidate_spec_id,
-            "candidate_id": candidate_id,
-            "hypothesis_id": hypothesis_id,
-            "observed_stage": "paper",
-            "strategy_family": strategy_family,
-            "strategy_name": strategy_name,
-            "account_label": account_label,
-            "window_start": window_start,
-            "window_end": window_end,
-            "source_kind": "simulation_exact_replay_runtime_ledger"
-            if exact_replay_ledger_artifact_refs
-            else "paper_runtime_observed",
-            "source_manifest_ref": _candidate_board_hypothesis_manifest_ref(
-                hypothesis_id
-            ),
-            "dataset_snapshot_ref": _string(row.get("dataset_snapshot_id")),
-            "artifact_refs": [
-                _string(ref)
-                for ref in cast(Sequence[Any], row.get("replay_artifact_refs") or ())
-                if _string(ref)
-            ],
-            "exact_replay_ledger_artifact_refs": list(
-                exact_replay_ledger_artifact_refs
-            ),
-            "exact_replay_ledger_artifact_ref": exact_replay_ledger_artifact_ref,
-            "exact_replay_ledger_artifact_row_count": int(
-                _decimal(row.get("exact_replay_ledger_artifact_row_count"))
-            ),
-            "exact_replay_ledger_artifact_fill_count": int(
-                _decimal(row.get("exact_replay_ledger_artifact_fill_count"))
-            ),
-            "candidate_blockers": [
-                _string(blocker)
-                for blocker in cast(Sequence[Any], row.get("blockers") or ())
-                if _string(blocker)
-            ],
-            "handoff": "runtime_window_import_only",
-            "promotion_gate": "existing_runtime_governance_fail_closed",
-        }
-        if runtime_ledger_lineage_handoff:
-            target["runtime_ledger_lineage_materialization_handoff"] = dict(
-                runtime_ledger_lineage_handoff
-            )
-            target["runtime_ledger_required_materialized_artifacts"] = (
-                runtime_ledger_required_materialized_artifacts
-            )
-            materialization_status = _string(
-                row.get("runtime_ledger_materialization_status")
-                or runtime_ledger_lineage_handoff.get("materialization_status")
-                or runtime_ledger_lineage_handoff.get("status")
-            )
-            if materialization_status:
-                target["runtime_ledger_materialization_status"] = materialization_status
-            if _boolish(
-                row.get("zero_authoritative_daily_pnl_until_materialized")
-                or runtime_ledger_lineage_handoff.get(
-                    "zero_authoritative_daily_pnl_until_materialized"
-                )
-            ):
-                target["zero_authoritative_daily_pnl_until_materialized"] = True
-        if bool(row.get("paper_probation_authorized")):
-            target.update(
-                {
-                    "candidate_selection": _string(row.get("candidate_selection")),
-                    "paper_probation_authorized": True,
-                    "paper_probation_authorization_scope": _string(
-                        row.get("paper_probation_authorization_scope")
-                    ),
-                    "evidence_collection_stage": _string(
-                        row.get("evidence_collection_stage")
-                    ),
-                    "probation_allowed": bool(row.get("probation_allowed")),
-                    "probation_reason": _string(row.get("probation_reason")),
-                    "selection_reason": _string(row.get("selection_reason")),
-                    "selected_by": _string(row.get("selected_by")),
-                    "probation_lower_bound_net_pnl_per_day": _string(
-                        row.get("probation_lower_bound_net_pnl_per_day")
-                    ),
-                    "probation_target_shortfall": _string(
-                        row.get("probation_target_shortfall")
-                    ),
-                    "probation_target_progress_ratio": _string(
-                        row.get("probation_target_progress_ratio")
-                    ),
-                    "required_notional_repair_scale_to_target": _string(
-                        row.get("required_notional_repair_scale_to_target")
-                    ),
-                    "required_notional_repair_scale_authority": _string(
-                        row.get("required_notional_repair_scale_authority")
-                    ),
-                    "live_paper_evidence_requirements": _string_list_from_value(
-                        row.get("live_paper_evidence_requirements")
-                    ),
-                    "safe_evidence_collection_path": _string_list_from_value(
-                        row.get("safe_evidence_collection_path")
-                    ),
-                    "live_capital_authorized": False,
-                    "final_promotion_requires_live_paper_runtime_proof": _boolish(
-                        row.get("final_promotion_requires_live_paper_runtime_proof")
-                    ),
-                    "promotion_allowed": False,
-                    "final_promotion_authorized": False,
-                    "final_promotion_allowed": False,
-                    "final_promotion_blockers": [
-                        _string(blocker)
-                        for blocker in cast(
-                            Sequence[Any], row.get("final_promotion_blockers") or ()
-                        )
-                        if _string(blocker)
-                    ],
-                }
-            )
-        paper_required_evidence_tokens = _string_list_from_value(
-            row.get("paper_required_evidence_tokens")
-        )
-        paper_mechanism_overlay_ids = _string_list_from_value(
-            row.get("paper_mechanism_overlay_ids")
-        )
-        paper_contract_prior_score = _string(row.get("paper_contract_prior_score"))
-        if (
-            _boolish(row.get("paper_contract_candidate"))
-            or paper_contract_prior_score
-            or paper_required_evidence_tokens
-            or paper_mechanism_overlay_ids
-        ):
-            target.update(
-                {
-                    "replay_selection_reason": _string(
-                        row.get("replay_selection_reason")
-                    ),
-                    "paper_contract_candidate": bool(
-                        row.get("paper_contract_candidate")
-                    ),
-                    "paper_contract_selected_for_replay": bool(
-                        row.get("paper_contract_selected_for_replay")
-                    ),
-                    "paper_contract_prior_score": paper_contract_prior_score,
-                    "paper_mechanism_overlay_ids": paper_mechanism_overlay_ids,
-                    "paper_required_evidence_tokens": paper_required_evidence_tokens,
-                    "paper_required_evidence_count": _candidate_board_int_field(
-                        row, "paper_required_evidence_count"
-                    ),
-                }
-            )
-        target["target_metadata"] = {
-            key: value
-            for key, value in target.items()
-            if key
-            not in {
-                "import_command_args",
-                "target_metadata",
-            }
-        }
-        target["import_command_args"] = _candidate_board_runtime_import_args(target)
-        targets.append(target)
-
-    return {
-        "schema_version": "torghut.runtime-window-import-plan.v1",
-        "status": "ready" if targets else "blocked",
-        "observed_stage": "paper",
-        "target_count": len(targets),
-        "targets": targets,
-        "blockers": blockers,
-    }
-
-
-def _candidate_factor_acceptance_replay_metadata(
-    *,
-    spec: CandidateSpec,
-    evidence: CandidateEvidenceBundle | None,
-    candidate_count: int,
-) -> dict[str, Any]:
-    static_artifact = _mapping(spec.feature_contract.get("factor_acceptance_artifact"))
-    if not static_artifact:
-        return {}
-
-    if evidence is None:
-        artifact = {
-            **static_artifact,
-            "status": "rejected",
-            "rejection_reasons": list(
-                dict.fromkeys(
-                    [
-                        *_string_list_from_value(
-                            static_artifact.get("rejection_reasons")
-                        ),
-                        "replay_evidence_missing",
-                    ]
-                )
-            ),
-            "promotion_scope": "research_paper_probation_only",
-            "does_not_authorize_live_promotion": True,
-        }
-        evidence_status = "missing"
-        candidate_id = ""
-        evidence_bundle_id = ""
-    else:
-        artifact = build_factor_acceptance_artifact_from_scorecard(
-            factor_expression=_string(static_artifact.get("factor_expression"))
-            or "candidate_family_score",
-            source_idea=_string(static_artifact.get("source_idea"))
-            or "2025_2026_signal_discovery_rankic_acceptance_harness",
-            allowed_feature_dependencies=_string_list_from_value(
-                static_artifact.get("allowed_feature_dependencies")
-            )
-            or _string_list_from_value(spec.feature_contract.get("required_features")),
-            scorecard=evidence.objective_scorecard,
-            fold_metrics=evidence.fold_metrics,
-            candidate_count=max(1, candidate_count),
-            candidate_spec_id=spec.candidate_spec_id,
-            candidate_id=evidence.candidate_id,
-            evidence_bundle_id=evidence.evidence_bundle_id,
-        )
-        evidence_status = "replayed"
-        candidate_id = evidence.candidate_id
-        evidence_bundle_id = evidence.evidence_bundle_id
-
-    return {
-        "schema_version": "torghut.factor-acceptance-replay-metadata.v1",
-        "candidate_spec_id": spec.candidate_spec_id,
-        "candidate_id": candidate_id,
-        "evidence_bundle_id": evidence_bundle_id,
-        "evidence_status": evidence_status,
-        "status": _string(artifact.get("status")) or "rejected",
-        "rejection_reasons": _string_list_from_value(artifact.get("rejection_reasons")),
-        "lineage_hash": _string(artifact.get("lineage_hash")),
-        "replay_artifact": artifact,
-        "static_compile_artifact_lineage_hash": _string(
-            static_artifact.get("lineage_hash")
-        ),
-        "promotion_scope": "research_paper_probation_only",
-        "promotion_allowed": False,
-        "final_promotion_authorized": False,
-        "does_not_authorize_live_promotion": True,
-    }
-
-
-def _candidate_board_factor_acceptance_summary(
-    rows: Sequence[Mapping[str, Any]],
-) -> dict[str, Any]:
-    metadata_rows = [
-        _mapping(row.get("factor_acceptance_replay_metadata"))
-        for row in rows
-        if _mapping(row.get("factor_acceptance_replay_metadata"))
-    ]
-    accepted = [
-        row for row in metadata_rows if _string(row.get("status")) == "accepted"
-    ]
-    return {
-        "schema_version": "torghut.factor-acceptance-board-summary.v1",
-        "candidate_count": len(metadata_rows),
-        "accepted_count": len(accepted),
-        "rejected_count": len(metadata_rows) - len(accepted),
-        "promotion_allowed": False,
-        "final_promotion_authorized": False,
-        "scope": "research_paper_probation_only",
-        "accepted_candidate_spec_ids": [
-            _string(row.get("candidate_spec_id")) for row in accepted
-        ],
-    }
-
-
-def _candidate_board_payload(
-    *,
-    epoch_id: str,
-    output_dir: Path,
-    target: Decimal,
-    candidate_specs: Sequence[CandidateSpec],
-    candidate_selection: Mapping[str, Any],
-    pre_replay_proposal_rows: Sequence[Mapping[str, Any]],
-    proposal_rows: Sequence[Mapping[str, Any]],
-    evidence_bundles: Sequence[CandidateEvidenceBundle],
-    portfolio: PortfolioCandidateSpec | None,
-    promotion_readiness: Mapping[str, Any],
-    runtime_closure: Mapping[str, Any],
-    paper_probation_target_limit: int = 1,
-) -> dict[str, Any]:
-    pre_replay_by_spec = _candidate_board_score_rows(pre_replay_proposal_rows)
-    proposal_by_spec = _candidate_board_score_rows(proposal_rows)
-    selection_by_spec = _candidate_board_score_rows(
-        _list_of_mappings(candidate_selection.get("rows"))
-    )
-    evidence_by_spec = {bundle.candidate_spec_id: bundle for bundle in evidence_bundles}
-    portfolio_payload = portfolio.to_payload() if portfolio is not None else {}
-    portfolio_scorecard = _mapping(portfolio_payload.get("objective_scorecard"))
-    portfolio_oracle_passed = _boolish(portfolio_scorecard.get("oracle_passed"))
-    portfolio_sleeve_spec_ids = {
-        _string(sleeve.get("candidate_spec_id"))
-        for sleeve in _list_of_mappings(portfolio_payload.get("sleeves"))
-        if _string(sleeve.get("candidate_spec_id"))
-    }
-    factor_acceptance_candidate_count = max(
-        _candidate_board_int_field(
-            _mapping(candidate_selection.get("budget")),
-            "compiled_candidate_count",
-        ),
-        len(candidate_specs),
-        len(evidence_bundles),
-        1,
-    )
-    rows: list[dict[str, Any]] = []
-    for spec in candidate_specs:
-        selection = selection_by_spec.get(spec.candidate_spec_id, {})
-        pre_replay = pre_replay_by_spec.get(spec.candidate_spec_id, {})
-        proposal = proposal_by_spec.get(spec.candidate_spec_id, pre_replay)
-        evidence = evidence_by_spec.get(spec.candidate_spec_id)
-        raw_scorecard = (
-            dict(evidence.objective_scorecard) if evidence is not None else {}
-        )
-        mechanism_overlay_ids = _string_list_from_value(
-            spec.parameter_space.get("mechanism_overlay_ids")
-        )
-        if mechanism_overlay_ids:
-            raw_scorecard.setdefault("mechanism_overlay_ids", mechanism_overlay_ids)
-        if evidence is not None:
-            raw_scorecard.setdefault(
-                "replay_artifact_refs", list(evidence.replay_artifact_refs)
-            )
-        scorecard = _candidate_board_scorecard_with_evidence_blockers(
-            raw_scorecard, evidence
-        )
-        scorecard, evidence_lineage = _candidate_board_scorecard_with_lineage_blockers(
-            scorecard, evidence
-        )
-        scorecard, replay_window_coverage = (
-            _candidate_board_scorecard_with_replay_window_blockers(scorecard)
-        )
-        market_impact_proof = _candidate_board_market_impact_proof_summary(scorecard)
-        regime_specialist_validation = _candidate_board_regime_specialist_summary(
-            spec, scorecard
-        )
-        scorecard, rejected_signal_summary = (
-            _candidate_board_scorecard_with_rejected_signal_blockers(spec, scorecard)
-        )
-        scorecard, order_type_summary = (
-            _candidate_board_scorecard_with_order_type_blockers(
-                spec,
-                scorecard,
-                replay_artifact_refs=evidence.replay_artifact_refs
-                if evidence is not None
-                else (),
-            )
-        )
-        scorecard, queue_position_survival_summary = (
-            _candidate_board_scorecard_with_queue_position_survival_blockers(
-                spec, scorecard
-            )
-        )
-        scorecard, predictability_decay_summary = (
-            _candidate_board_scorecard_with_predictability_decay_blockers(
-                spec, scorecard, target=target
-            )
-        )
-        selected_for_replay = bool(selection.get("selected_for_replay"))
-        replay_selection_reason = _string(selection.get("selection_reason")) or (
-            "not_selected_budget"
-        )
-        paper_contract_prior_score = _string(
-            selection.get("paper_contract_prior_score")
-        )
-        paper_mechanism_overlay_ids = _string_list_from_value(
-            selection.get("paper_mechanism_overlay_ids")
-        )
-        if not paper_mechanism_overlay_ids:
-            paper_mechanism_overlay_ids = mechanism_overlay_ids
-        paper_required_evidence_tokens = _string_list_from_value(
-            selection.get("paper_required_evidence_tokens")
-        )
-        paper_required_evidence_count = _candidate_board_int_field(
-            selection, "paper_required_evidence_count"
-        )
-        if paper_required_evidence_tokens and paper_required_evidence_count == 0:
-            paper_required_evidence_count = len(paper_required_evidence_tokens)
-        paper_contract_candidate = bool(
-            _decimal(paper_contract_prior_score) > 0
-            or paper_required_evidence_tokens
-            or paper_mechanism_overlay_ids
-        )
-        in_best_portfolio = spec.candidate_spec_id in portfolio_sleeve_spec_ids
-        blockers = _candidate_board_blockers(
-            selected_for_replay=selected_for_replay,
-            evidence=evidence,
-            scorecard=scorecard,
-        )
-        runtime_window_start, runtime_window_end = (
-            _candidate_board_runtime_window_bounds(scorecard)
-        )
-        exact_replay_ledger_artifact_ref = _string(
-            scorecard.get("exact_replay_ledger_artifact_ref")
-        )
-        runtime_ledger_lineage_handoff = (
-            _candidate_board_runtime_ledger_lineage_handoff(
-                scorecard=scorecard,
-                evidence=evidence,
-            )
-        )
-        runtime_ledger_required_materialized_artifacts = (
-            _candidate_board_runtime_ledger_required_materialized_artifacts(
-                runtime_ledger_lineage_handoff
-            )
-        )
-        runtime_ledger_materialization_status = _string(
-            runtime_ledger_lineage_handoff.get("materialization_status")
-            or runtime_ledger_lineage_handoff.get("status")
-        )
-        factor_acceptance_replay_metadata = (
-            _candidate_factor_acceptance_replay_metadata(
-                spec=spec,
-                evidence=evidence,
-                candidate_count=factor_acceptance_candidate_count,
-            )
-        )
-        rows.append(
-            {
-                "epoch_id": epoch_id,
-                "candidate_spec_id": spec.candidate_spec_id,
-                "candidate_id": evidence.candidate_id if evidence is not None else "",
-                "hypothesis_id": spec.hypothesis_id,
-                "family_template_id": spec.family_template_id,
-                "runtime_family": spec.runtime_family,
-                "runtime_strategy_name": spec.runtime_strategy_name,
-                "mechanism_overlay_ids": mechanism_overlay_ids,
-                "pre_replay_rank": _rank_sort_value(pre_replay.get("rank")),
-                "pre_replay_proposal_score": str(
-                    pre_replay.get("proposal_score") or ""
-                ),
-                "rank": _rank_sort_value(proposal.get("rank")),
-                "proposal_score": str(proposal.get("proposal_score") or ""),
-                "selected_for_replay": selected_for_replay,
-                "replay_selection_reason": replay_selection_reason,
-                "paper_contract_candidate": paper_contract_candidate,
-                "paper_contract_selected_for_replay": bool(
-                    selected_for_replay and paper_contract_candidate
-                ),
-                "paper_contract_prior_score": paper_contract_prior_score,
-                "paper_mechanism_overlay_ids": paper_mechanism_overlay_ids,
-                "paper_required_evidence_tokens": paper_required_evidence_tokens,
-                "paper_required_evidence_count": paper_required_evidence_count,
-                "has_replay_evidence": evidence is not None,
-                "in_best_portfolio": in_best_portfolio,
-                "dataset_snapshot_id": evidence.dataset_snapshot_id
-                if evidence is not None
-                else "",
-                "status": _candidate_board_status(
-                    selected_for_replay=selected_for_replay,
-                    evidence=evidence,
-                    scorecard=scorecard,
-                    in_best_portfolio=in_best_portfolio,
-                    portfolio_oracle_passed=portfolio_oracle_passed,
-                ),
-                "target_met": _boolish(scorecard.get("target_met")),
-                "oracle_passed": _boolish(scorecard.get("oracle_passed")),
-                "net_pnl_per_day": _candidate_board_decimal_field(
-                    scorecard, "net_pnl_per_day"
-                )
-                or _candidate_board_decimal_field(
-                    scorecard, "portfolio_post_cost_net_pnl_per_day"
-                ),
-                "market_impact_stress_passed": _boolish(
-                    scorecard.get("market_impact_stress_passed")
-                ),
-                "market_impact_stress_net_pnl_per_day": _candidate_board_decimal_field(
-                    scorecard, "market_impact_stress_net_pnl_per_day"
-                ),
-                "delay_adjusted_depth_stress_passed": _boolish(
-                    scorecard.get("delay_adjusted_depth_stress_passed")
-                ),
-                "delay_adjusted_depth_stress_net_pnl_per_day": _candidate_board_decimal_field(
-                    scorecard, "delay_adjusted_depth_stress_net_pnl_per_day"
-                ),
-                "delay_adjusted_depth_fill_survival_evidence_present": _boolish(
-                    scorecard.get("delay_adjusted_depth_fill_survival_evidence_present")
-                    or scorecard.get("fill_survival_evidence_present")
-                ),
-                "delay_adjusted_depth_fill_survival_sample_count": _candidate_board_first_int_field(
-                    scorecard,
-                    (
-                        "delay_adjusted_depth_fill_survival_sample_count",
-                        "fill_survival_sample_count",
-                    ),
-                ),
-                "delay_adjusted_depth_fill_survival_rate": _candidate_board_decimal_field(
-                    scorecard, "delay_adjusted_depth_fill_survival_rate"
-                )
-                or _candidate_board_decimal_field(scorecard, "fill_survival_fill_rate")
-                or _candidate_board_decimal_field(scorecard, "fill_survival_rate"),
-                "queue_position_survival_fill_curve_evidence_present": _boolish(
-                    scorecard.get("queue_position_survival_fill_curve_evidence_present")
-                ),
-                "queue_position_survival_sample_count": _candidate_board_int_field(
-                    scorecard, "queue_position_survival_sample_count"
-                ),
-                "queue_position_survival_fill_rate": _candidate_board_decimal_field(
-                    scorecard, "queue_position_survival_fill_rate"
-                ),
-                "queue_position_survival_queue_ratio_p95": _candidate_board_decimal_field(
-                    scorecard, "queue_position_survival_queue_ratio_p95"
-                ),
-                "queue_position_survival_queue_ahead_depletion_evidence_present": _boolish(
-                    scorecard.get(
-                        "queue_position_survival_queue_ahead_depletion_evidence_present"
-                    )
-                ),
-                "queue_position_survival_queue_ahead_depletion_sample_count": (
-                    _candidate_board_int_field(
-                        scorecard,
-                        "queue_position_survival_queue_ahead_depletion_sample_count",
-                    )
-                ),
-                "queue_position_survival_nonfill_opportunity_cost_bps": (
-                    _candidate_board_decimal_field(
-                        scorecard,
-                        "queue_position_survival_nonfill_opportunity_cost_bps",
-                    )
-                    or _candidate_board_decimal_field(
-                        scorecard, "queue_position_nonfill_opportunity_cost_bps"
-                    )
-                ),
-                "fill_time_ms_p50": _candidate_board_decimal_field(
-                    scorecard, "fill_time_ms_p50"
-                ),
-                "fill_time_ms_p95": _candidate_board_decimal_field(
-                    scorecard, "fill_time_ms_p95"
-                ),
-                "delay_adjusted_depth_queue_ahead_depletion_evidence_present": _boolish(
-                    scorecard.get(
-                        "delay_adjusted_depth_queue_ahead_depletion_evidence_present"
-                    )
-                ),
-                "delay_adjusted_depth_queue_ahead_depletion_sample_count": (
-                    _candidate_board_int_field(
-                        scorecard,
-                        "delay_adjusted_depth_queue_ahead_depletion_sample_count",
-                    )
-                ),
-                "queue_ahead_depletion_evidence_present": _boolish(
-                    scorecard.get("queue_ahead_depletion_evidence_present")
-                ),
-                "queue_ahead_depletion_sample_count": _candidate_board_int_field(
-                    scorecard, "queue_ahead_depletion_sample_count"
-                ),
-                "post_cost_net_pnl_after_queue_position_survival_fill_stress": (
-                    _candidate_board_decimal_field(
-                        scorecard,
-                        "post_cost_net_pnl_after_queue_position_survival_fill_stress",
-                    )
-                ),
-                "implementation_uncertainty_stability_passed": _boolish(
-                    scorecard.get("implementation_uncertainty_stability_passed")
-                ),
-                "implementation_uncertainty_lower_net_pnl_per_day": _candidate_board_decimal_field(
-                    scorecard, "implementation_uncertainty_lower_net_pnl_per_day"
-                ),
-                "conformal_tail_risk_required": _boolish(
-                    scorecard.get("conformal_tail_risk_required")
-                ),
-                "conformal_tail_risk_passed": _boolish(
-                    scorecard.get("conformal_tail_risk_passed")
-                ),
-                "conformal_tail_risk_adjusted_net_pnl_per_day": _candidate_board_decimal_field(
-                    scorecard, "conformal_tail_risk_adjusted_net_pnl_per_day"
-                ),
-                "conformal_tail_risk_buffer_per_day": _candidate_board_decimal_field(
-                    scorecard, "conformal_tail_risk_buffer_per_day"
-                ),
-                "breakeven_transaction_cost_buffer_passed": _boolish(
-                    scorecard.get("breakeven_transaction_cost_buffer_passed")
-                ),
-                "breakeven_transaction_cost_buffer_bps": _candidate_board_decimal_field(
-                    scorecard, "breakeven_transaction_cost_buffer_bps"
-                ),
-                "transaction_cost_buffer_bps": _candidate_board_decimal_field(
-                    scorecard, "transaction_cost_buffer_bps"
-                ),
-                "post_cost_net_pnl_after_breakeven_transaction_cost_buffer": (
-                    _candidate_board_decimal_field(
-                        scorecard,
-                        "post_cost_net_pnl_after_breakeven_transaction_cost_buffer",
-                    )
-                ),
-                "required_seed_model_family_robustness": _boolish(
-                    scorecard.get("required_seed_model_family_robustness")
-                ),
-                "seed_model_family_robustness_status": str(
-                    scorecard.get("seed_model_family_robustness_status") or ""
-                ),
-                "seed_robustness_passed": _boolish(
-                    scorecard.get("seed_robustness_passed")
-                ),
-                "seed_robustness_sample_count": _candidate_board_int_field(
-                    scorecard, "seed_robustness_sample_count"
-                ),
-                "model_family_robustness_passed": _boolish(
-                    scorecard.get("model_family_robustness_passed")
-                ),
-                "model_family_robustness_family_count": _candidate_board_int_field(
-                    scorecard, "model_family_robustness_family_count"
-                ),
-                "trading_day_count": _candidate_board_int_field(
-                    scorecard, "trading_day_count"
-                ),
-                "decision_count": _candidate_board_first_int_field(
-                    scorecard,
-                    (
-                        "decision_count",
-                        "trade_decision_count",
-                        "paper_decision_count",
-                        "runtime_decision_count",
-                    ),
-                ),
-                "submitted_order_count": _candidate_board_first_int_field(
-                    scorecard,
-                    (
-                        "submitted_order_count",
-                        "order_count",
-                        "orders_submitted_count",
-                        "paper_order_count",
-                        "runtime_order_count",
-                    ),
-                ),
-                "filled_order_count": _candidate_board_first_int_field(
-                    scorecard,
-                    (
-                        "filled_count",
-                        "filled_order_count",
-                        "executions_filled_count",
-                        "execution_filled_count",
-                        "paper_filled_order_count",
-                        "trade_count",
-                        "runtime_trade_count",
-                    ),
-                ),
-                "executable_replay_order_count": _candidate_board_first_int_field(
-                    scorecard,
-                    (
-                        "executable_replay_order_count",
-                        "executable_replay_submitted_order_count",
-                        "executable_replay_orders_submitted_total",
-                    ),
-                ),
-                "double_oos_passed": _boolish(scorecard.get("double_oos_passed")),
-                "double_oos_artifact_ref": _string(
-                    scorecard.get("double_oos_artifact_ref")
-                    or scorecard.get("double_oos_report_ref")
-                    or scorecard.get("walk_forward_oos_artifact_ref")
-                ),
-                "double_oos_independent_window_count": _candidate_board_first_int_field(
-                    scorecard,
-                    (
-                        "double_oos_independent_window_count",
-                        "double_oos_fold_count",
-                        "oos_fold_count",
-                    ),
-                ),
-                "double_oos_pass_rate": _candidate_board_decimal_field(
-                    scorecard, "double_oos_pass_rate"
-                ),
-                "double_oos_net_pnl_per_day": _candidate_board_decimal_field(
-                    scorecard, "double_oos_net_pnl_per_day"
-                ),
-                "double_oos_cost_shock_net_pnl_per_day": _candidate_board_decimal_field(
-                    scorecard, "double_oos_cost_shock_net_pnl_per_day"
-                ),
-                "active_day_ratio": _candidate_board_decimal_field(
-                    scorecard, "active_day_ratio"
-                ),
-                "positive_day_ratio": _candidate_board_decimal_field(
-                    scorecard, "positive_day_ratio"
-                ),
-                "best_day_share": _candidate_board_decimal_field(
-                    scorecard, "best_day_share"
-                ),
-                "worst_day_loss": _candidate_board_decimal_field(
-                    scorecard, "worst_day_loss"
-                ),
-                "avg_filled_notional_per_day": _candidate_board_decimal_field(
-                    scorecard, "avg_filled_notional_per_day"
-                ),
-                "market_impact_proof": market_impact_proof,
-                "regime_specialist_validation": regime_specialist_validation,
-                "rejected_signal_outcome_learning": rejected_signal_summary,
-                "order_type_execution_quality": order_type_summary,
-                "queue_position_survival_fill_quality": (
-                    queue_position_survival_summary
-                ),
-                "predictability_decay_stress": predictability_decay_summary,
-                "factor_acceptance_replay_metadata": (
-                    factor_acceptance_replay_metadata
-                ),
-                "factor_acceptance_status": _string(
-                    factor_acceptance_replay_metadata.get("status")
-                ),
-                "factor_acceptance_promotion_allowed": False,
-                "evidence_lineage": evidence_lineage,
-                "replay_window_coverage": replay_window_coverage,
-                "blockers": blockers,
-                "runtime_window_start": runtime_window_start,
-                "runtime_window_end": runtime_window_end,
-                "account_label": _string(scorecard.get("account_label"))
-                or ("TORGHUT_REPLAY" if exact_replay_ledger_artifact_ref else ""),
-                "exact_replay_ledger_artifact_ref": exact_replay_ledger_artifact_ref,
-                "exact_replay_ledger_artifact_row_count": (
-                    _candidate_board_first_int_field(
-                        scorecard,
-                        ("exact_replay_ledger_artifact_row_count",),
-                    )
-                ),
-                "exact_replay_ledger_artifact_fill_count": (
-                    _candidate_board_first_int_field(
-                        scorecard,
-                        ("exact_replay_ledger_artifact_fill_count",),
-                    )
-                ),
-                "runtime_ledger_lineage_materialization_handoff": dict(
-                    runtime_ledger_lineage_handoff
-                ),
-                "runtime_ledger_required_materialized_artifacts": (
-                    runtime_ledger_required_materialized_artifacts
-                ),
-                "runtime_ledger_materialization_status": (
-                    runtime_ledger_materialization_status
-                ),
-                "zero_authoritative_daily_pnl_until_materialized": _boolish(
-                    runtime_ledger_lineage_handoff.get(
-                        "zero_authoritative_daily_pnl_until_materialized"
-                    )
-                ),
-                "replay_artifact_refs": list(evidence.replay_artifact_refs)
-                if evidence is not None
-                else [],
-            }
-        )
-    rows.sort(
-        key=lambda row: (
-            _rank_sort_value(row.get("rank")),
-            -_proposal_sort_value(row.get("proposal_score")),
-            _string(row.get("candidate_spec_id")),
-        )
-    )
-    best_research_candidate = rows[0] if rows else None
-    best_executed_candidate = _candidate_board_best_executed_candidate(rows)
-    closest_promotion_candidate = _candidate_board_closest_promotion_candidate(rows)
-    paper_probation_candidates = _candidate_board_paper_probation_candidates(
-        rows,
-        target=target,
-        limit=paper_probation_target_limit,
-    )
-    paper_probation_candidate = (
-        paper_probation_candidates[0] if paper_probation_candidates else None
-    )
-    promotion_ready = _boolish(promotion_readiness.get("promotable"))
-    promotion_subject = _candidate_board_portfolio_promotion_subject(
-        portfolio=portfolio,
-        portfolio_payload=portfolio_payload,
-        portfolio_scorecard=portfolio_scorecard,
-        promotion_readiness=promotion_readiness,
-        runtime_closure=runtime_closure,
-        rows=rows,
-    )
-    promotion_candidate_found = (
-        promotion_ready and portfolio_oracle_passed and promotion_subject is not None
-    ) or (
-        promotion_ready
-        and closest_promotion_candidate is not None
-        and bool(closest_promotion_candidate.get("oracle_passed"))
-    )
-    return {
-        "schema_version": "torghut.profit-candidate-board.v1",
-        "epoch_id": epoch_id,
-        "run_root": str(output_dir),
-        "target_net_pnl_per_day": str(target),
-        "status_digest": _candidate_board_status_digest(rows),
-        "current_answer": "promotion_candidate_found"
-        if promotion_candidate_found
-        else "no_promotion_ready_candidate",
-        "promotion_readiness": dict(promotion_readiness),
-        "best_research_candidate": best_research_candidate,
-        "best_executed_candidate": best_executed_candidate,
-        "closest_promotion_candidate": closest_promotion_candidate,
-        "paper_probation_candidate": paper_probation_candidate,
-        "paper_probation_candidates": list(paper_probation_candidates),
-        "paper_probation_target_limit": max(1, int(paper_probation_target_limit)),
-        "promotion_subject": promotion_subject,
-        "runtime_window_import_plan": _candidate_board_runtime_window_import_plan(
-            rows=rows,
-            paper_probation_candidate=None,
-            paper_probation_candidates=paper_probation_candidates,
-            promotion_subject=promotion_subject,
-        ),
-        "factor_acceptance_summary": _candidate_board_factor_acceptance_summary(rows),
-        "double_oos_summary": _candidate_board_double_oos_summary(rows),
-        "best_portfolio_candidate_id": _string(
-            portfolio_payload.get("portfolio_candidate_id")
-        ),
-        "best_portfolio_oracle_passed": portfolio_oracle_passed,
-        "runtime_closure_status": _string(runtime_closure.get("status")),
-        "row_count": len(rows),
-        "rows": rows,
-    }
-
-
-def _paper_probation_handoff_payload(
-    candidate_board: Mapping[str, Any],
-) -> dict[str, Any]:
-    candidates = _list_of_mappings(candidate_board.get("paper_probation_candidates"))
-    import_plan = _mapping(candidate_board.get("runtime_window_import_plan"))
-    import_plan_ready = _string(import_plan.get("status")) == "ready"
-    blockers: list[str] = []
-    if not candidates:
-        blockers.append("paper_probation_candidate_missing")
-    if not import_plan_ready:
-        blockers.append("runtime_window_import_plan_not_ready")
-    status = "ready" if candidates and import_plan_ready else "not_ready"
-    return {
-        "schema_version": "torghut.paper-probation-handoff.v1",
-        "status": status,
-        "current_answer": _string(candidate_board.get("current_answer")),
-        "evidence_collection_stage": "paper",
-        "authorization_scope": "evidence_collection_only",
-        "paper_probation_authorized": status == "ready",
-        "probation_allowed": status == "ready",
-        "promotion_allowed": False,
-        "final_promotion_authorized": False,
-        "final_promotion_allowed": False,
-        "promotion_gate": "existing_runtime_governance_fail_closed",
-        "blockers": blockers,
-        "candidate_count": len(candidates),
-        "candidates": candidates,
-        "runtime_window_import_plan": import_plan,
-    }
-
-
-def _portfolio_with_runtime_closure_proof(
-    *,
-    portfolio: PortfolioCandidateSpec,
-    runtime_closure: Mapping[str, Any],
-    target: Decimal,
-    oracle_policy: ProfitTargetOraclePolicy,
-) -> PortfolioCandidateSpec:
-    proof_update = _runtime_closure_scorecard_update(
-        portfolio=portfolio, runtime_closure=runtime_closure
-    )
-    proof_payload = _mapping(proof_update.get("runtime_closure_proof"))
-    if not proof_payload.get("executable_replay_artifact_refs") and proof_payload.get(
-        "shadow_status"
-    ) not in {"within_budget", "invalid_artifact"}:
-        return portfolio
-
-    scorecard = {**dict(portfolio.objective_scorecard), **proof_update}
-    scorecard["profit_target_oracle"] = evaluate_profit_target_oracle(
-        scorecard,
-        target_net_pnl_per_day=target,
-        policy=oracle_policy,
-    )
-    scorecard["oracle_passed"] = bool(scorecard["profit_target_oracle"]["passed"])
-    runtime_refs = tuple(
-        ref
-        for ref in cast(
-            Sequence[Any], scorecard.get("runtime_closure_artifact_refs") or ()
-        )
-        if _string(ref)
-    )
-    evidence_refs = tuple(dict.fromkeys((*portfolio.evidence_refs, *runtime_refs)))
-    optimizer_report = {
-        **dict(portfolio.optimizer_report),
-        "runtime_closure_proof_status": proof_payload.get("status"),
-        "runtime_closure_artifact_count": len(runtime_refs),
-        "oracle_passed_after_runtime_closure": scorecard["oracle_passed"],
-    }
-    return replace(
-        portfolio,
-        objective_scorecard=scorecard,
-        evidence_refs=evidence_refs,
-        optimizer_report=optimizer_report,
-    )
-
-
-def _runtime_closure_program_for_candidate(
-    *,
-    program: StrategyAutoresearchProgram,
-    manifest: MlxSnapshotManifest,
-    portfolio: PortfolioCandidateSpec | None,
-    oracle_candidate_found: bool,
-) -> StrategyAutoresearchProgram:
-    runtime_window_available = bool(manifest.source_window_start) and bool(
-        manifest.source_window_end
-    )
-    if portfolio is None:
-        return program
-    if runtime_window_available and (
-        oracle_candidate_found or _portfolio_needs_runtime_closure_proof(portfolio)
-    ):
-        return program
-    return replace(
-        program,
-        runtime_closure_policy=replace(
-            program.runtime_closure_policy,
-            execute_parity_replay=False,
-            execute_approval_replay=False,
-        ),
-    )
 
 
 def run_whitepaper_autoresearch_profit_target(
