+++ services/torghut/scripts/whitepaper_autoresearch_runner/__init__.py
@@ -1,0 +1,1 @@
+"""Explicit owner modules for the whitepaper autoresearch profit target runner."""
