+++ services/torghut/tests/autoresearch_runner/test_runtime_closure.py
@@ -155,7 +155,10 @@ class TestAutoresearchRunnerRuntimeClosure(AutoresearchRunnerTestCase):
                 rows=_authoritative_exact_replay_ledger_rows(),
             )
         )
-        with patch.object(runner, "build_runtime_ledger_buckets", return_value=[]):
+        with patch(
+            "scripts.whitepaper_autoresearch_runner.runtime_closure.build_runtime_ledger_buckets",
+            return_value=[],
+        ):
             self.assertIsNone(
                 runner._runtime_closure_exact_replay_bucket(
                     ledger={
